@@ -103,7 +103,10 @@ class AdaptiveScaffold extends StatefulWidget {
     this.navigationRailWidth = 72,
     this.extendedNavigationRailWidth = 192,
     this.appBarBreakpoint,
-  });
+  }) : assert(
+          destinations.length >= 2,
+          'At least two destinations are required',
+        );
 
   /// The destinations to be used in navigation items. These are converted to
   /// [NavigationRailDestination]s and [BottomNavigationBarItem]s and inserted
@@ -502,12 +505,16 @@ class AdaptiveScaffold extends StatefulWidget {
 }
 
 class _AdaptiveScaffoldState extends State<AdaptiveScaffold> {
+  // Global scaffold key that will help to manage drawer state.
+  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
+
   @override
   Widget build(BuildContext context) {
     final NavigationRailThemeData navRailTheme =
         Theme.of(context).navigationRailTheme;
 
     return Scaffold(
+      key: _scaffoldKey,
       appBar: widget.drawerBreakpoint.isActive(context) && widget.useDrawer ||
               (widget.appBarBreakpoint?.isActive(context) ?? false)
           ? widget.appBar ?? AppBar()
@@ -523,7 +530,7 @@ class _AdaptiveScaffoldState extends State<AdaptiveScaffold> {
                     .map((NavigationDestination destination) =>
                         AdaptiveScaffold.toRailDestination(destination))
                     .toList(),
-                onDestinationSelected: widget.onSelectedIndexChange,
+                onDestinationSelected: _onDrawerDestinationSelected,
               ),
             )
           : null,
@@ -669,6 +676,20 @@ class _AdaptiveScaffoldState extends State<AdaptiveScaffold> {
         ),
       ),
     );
+  }
+
+  void _onDrawerDestinationSelected(int index) {
+    if (widget.useDrawer) {
+      // If [useDrawer] is true, then retrieve the current state.
+      final ScaffoldState? scaffoldCurrentContext = _scaffoldKey.currentState;
+      if (scaffoldCurrentContext != null) {
+        if (scaffoldCurrentContext.isDrawerOpen) {
+          // If drawer is open, call [closeDrawer] to dismiss drawer as per material guidelines.
+          scaffoldCurrentContext.closeDrawer();
+        }
+      }
+    }
+    widget.onSelectedIndexChange?.call(index);
   }
 }
 
