@@ -177,6 +177,12 @@ class SwiftFunction {
   final String value;
 }
 
+/// Metadata to annotate data classes to be defined as class in Swift output.
+class SwiftClass {
+  /// Constructor.
+  const SwiftClass();
+}
+
 /// Type of TaskQueue which determines how handlers are dispatched for
 /// HostApi's.
 enum TaskQueueType {
@@ -1030,10 +1036,10 @@ List<Error> _validateProxyApi(
     }
 
     // Validate this api isn't used as an interface and contains anything except
-    // Flutter methods.
-    final bool isValidInterfaceProxyApi = api.hostMethods.isEmpty &&
-        api.constructors.isEmpty &&
-        api.fields.isEmpty;
+    // Flutter methods, a static host method, attached methods.
+    final bool isValidInterfaceProxyApi = api.constructors.isEmpty &&
+        api.fields.where((ApiField field) => !field.isStatic).isEmpty &&
+        api.hostMethods.where((Method method) => !method.isStatic).isEmpty;
     if (!isValidInterfaceProxyApi) {
       final Iterable<String> interfaceNames = proxyApi.interfaces.map(
         (TypeDeclaration type) => type.baseName,
@@ -1550,6 +1556,7 @@ class _RootBuilder extends dart_ast_visitor.RecursiveAstVisitor<Object?> {
       _currentClass = Class(
         name: node.name.lexeme,
         fields: <NamedType>[],
+        isSwiftClass: _hasMetadata(node.metadata, 'SwiftClass'),
         documentationComments:
             _documentationCommentsParser(node.documentationComment?.tokens),
       );
