@@ -106,6 +106,62 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
         allNullableTypesTwo.aNullableObject);
     expect(
         allNullableTypesOne.aNullableEnum, allNullableTypesTwo.aNullableEnum);
+    compareAllNullableTypes(allNullableTypesOne.allNullableTypes,
+        allNullableTypesTwo.allNullableTypes);
+  }
+
+  void compareAllNullableTypesWithoutRecursion(
+      AllNullableTypesWithoutRecursion? allNullableTypesOne,
+      AllNullableTypesWithoutRecursion? allNullableTypesTwo) {
+    expect(allNullableTypesOne == null, allNullableTypesTwo == null);
+    if (allNullableTypesOne == null || allNullableTypesTwo == null) {
+      return;
+    }
+    expect(
+        allNullableTypesOne.aNullableBool, allNullableTypesTwo.aNullableBool);
+    expect(allNullableTypesOne.aNullableInt, allNullableTypesTwo.aNullableInt);
+    expect(
+        allNullableTypesOne.aNullableInt64, allNullableTypesTwo.aNullableInt64);
+    expect(allNullableTypesOne.aNullableDouble,
+        allNullableTypesTwo.aNullableDouble);
+    expect(allNullableTypesOne.aNullableString,
+        allNullableTypesTwo.aNullableString);
+    expect(allNullableTypesOne.aNullableByteArray,
+        allNullableTypesTwo.aNullableByteArray);
+    expect(allNullableTypesOne.aNullable4ByteArray,
+        allNullableTypesTwo.aNullable4ByteArray);
+    expect(allNullableTypesOne.aNullable8ByteArray,
+        allNullableTypesTwo.aNullable8ByteArray);
+    expect(allNullableTypesOne.aNullableFloatArray,
+        allNullableTypesTwo.aNullableFloatArray);
+    expect(
+        listEquals(allNullableTypesOne.aNullableList,
+            allNullableTypesTwo.aNullableList),
+        true);
+    expect(
+        mapEquals(
+            allNullableTypesOne.aNullableMap, allNullableTypesTwo.aNullableMap),
+        true);
+    expect(allNullableTypesOne.nullableNestedList?.length,
+        allNullableTypesTwo.nullableNestedList?.length);
+    // TODO(stuartmorgan): Enable this once the Dart types are fixed; see
+    // https://github.com/flutter/flutter/issues/116117
+    //for (int i = 0; i < allNullableTypesOne.nullableNestedList!.length; i++) {
+    //  expect(listEquals(allNullableTypesOne.nullableNestedList![i], allNullableTypesTwo.nullableNestedList![i]),
+    //      true);
+    //}
+    expect(
+        mapEquals(allNullableTypesOne.nullableMapWithAnnotations,
+            allNullableTypesTwo.nullableMapWithAnnotations),
+        true);
+    expect(
+        mapEquals(allNullableTypesOne.nullableMapWithObject,
+            allNullableTypesTwo.nullableMapWithObject),
+        true);
+    expect(allNullableTypesOne.aNullableObject,
+        allNullableTypesTwo.aNullableObject);
+    expect(
+        allNullableTypesOne.aNullableEnum, allNullableTypesTwo.aNullableEnum);
   }
 
   void compareAllClassesWrapper(
@@ -117,6 +173,10 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
 
     compareAllNullableTypes(
         wrapperOne.allNullableTypes, wrapperTwo.allNullableTypes);
+    compareAllNullableTypesWithoutRecursion(
+      wrapperOne.allNullableTypesWithoutRecursion,
+      wrapperTwo.allNullableTypesWithoutRecursion,
+    );
     compareAllTypes(wrapperOne.allTypes, wrapperTwo.allTypes);
   }
 
@@ -170,6 +230,65 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
     aNullableObject: 0,
   );
 
+  final AllNullableTypes recursiveAllNullableTypes = AllNullableTypes(
+    aNullableBool: true,
+    aNullableInt: _regularInt,
+    aNullableInt64: _biggerThanBigInt,
+    aNullableDouble: _doublePi,
+    aNullableString: 'Hello host!',
+    aNullableByteArray: Uint8List.fromList(<int>[1, 2, 3]),
+    aNullable4ByteArray: Int32List.fromList(<int>[4, 5, 6]),
+    aNullable8ByteArray: Int64List.fromList(<int>[7, 8, 9]),
+    aNullableFloatArray: Float64List.fromList(<double>[2.71828, _doublePi]),
+    aNullableList: <Object?>['Thing 1', 2, true, 3.14, null],
+    aNullableMap: <Object?, Object?>{
+      'a': 1,
+      'b': 2.0,
+      'c': 'three',
+      'd': false,
+      'e': null
+    },
+    nullableNestedList: <List<bool>>[
+      <bool>[true, false],
+      <bool>[false, true]
+    ],
+    nullableMapWithAnnotations: <String?, String?>{},
+    nullableMapWithObject: <String?, Object?>{},
+    aNullableEnum: AnEnum.fourHundredTwentyTwo,
+    aNullableObject: 0,
+    allNullableTypes: genericAllNullableTypes,
+  );
+
+  final AllNullableTypesWithoutRecursion
+      genericAllNullableTypesWithoutRecursion =
+      AllNullableTypesWithoutRecursion(
+    aNullableBool: true,
+    aNullableInt: _regularInt,
+    aNullableInt64: _biggerThanBigInt,
+    aNullableDouble: _doublePi,
+    aNullableString: 'Hello host!',
+    aNullableByteArray: Uint8List.fromList(<int>[1, 2, 3]),
+    aNullable4ByteArray: Int32List.fromList(<int>[4, 5, 6]),
+    aNullable8ByteArray: Int64List.fromList(<int>[7, 8, 9]),
+    aNullableFloatArray: Float64List.fromList(<double>[2.71828, _doublePi]),
+    aNullableList: <Object?>['Thing 1', 2, true, 3.14, null],
+    aNullableMap: <Object?, Object?>{
+      'a': 1,
+      'b': 2.0,
+      'c': 'three',
+      'd': false,
+      'e': null
+    },
+    nullableNestedList: <List<bool>>[
+      <bool>[true, false],
+      <bool>[false, true]
+    ],
+    nullableMapWithAnnotations: <String?, String?>{},
+    nullableMapWithObject: <String?, Object?>{},
+    aNullableEnum: AnEnum.fourHundredTwentyTwo,
+    aNullableObject: 0,
+  );
+
   group('Host sync API tests', () {
     testWidgets('basic void->void call works', (WidgetTester _) async {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
@@ -190,9 +309,9 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
 
       final AllNullableTypes? echoObject =
-          await api.echoAllNullableTypes(genericAllNullableTypes);
+          await api.echoAllNullableTypes(recursiveAllNullableTypes);
 
-      compareAllNullableTypes(echoObject, genericAllNullableTypes);
+      compareAllNullableTypes(echoObject, recursiveAllNullableTypes);
     });
 
     testWidgets('all null datatypes serialize and deserialize correctly',
@@ -232,6 +351,67 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       compareAllNullableTypes(nullableListTypes, echoNullFilledClass);
     });
 
+    testWidgets(
+        'all nullable datatypes without recursion serialize and deserialize correctly',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion? echoObject =
+          await api.echoAllNullableTypesWithoutRecursion(
+              genericAllNullableTypesWithoutRecursion);
+
+      compareAllNullableTypesWithoutRecursion(
+          echoObject, genericAllNullableTypesWithoutRecursion);
+    });
+
+    testWidgets(
+        'all null datatypes without recursion serialize and deserialize correctly',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion allTypesNull =
+          AllNullableTypesWithoutRecursion();
+
+      final AllNullableTypesWithoutRecursion? echoNullFilledClass =
+          await api.echoAllNullableTypesWithoutRecursion(allTypesNull);
+      compareAllNullableTypesWithoutRecursion(
+          allTypesNull, echoNullFilledClass);
+    });
+
+    testWidgets(
+        'Classes without recursion with list of null serialize and deserialize correctly',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion nullableListTypes =
+          AllNullableTypesWithoutRecursion(
+              aNullableList: <String?>['String', null]);
+
+      final AllNullableTypesWithoutRecursion? echoNullFilledClass =
+          await api.echoAllNullableTypesWithoutRecursion(nullableListTypes);
+
+      compareAllNullableTypesWithoutRecursion(
+          nullableListTypes, echoNullFilledClass);
+    });
+
+    testWidgets(
+        'Classes without recursion with map of null serialize and deserialize correctly',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion nullableListTypes =
+          AllNullableTypesWithoutRecursion(aNullableMap: <String?, String?>{
+        'String': 'string',
+        'null': null
+      });
+
+      final AllNullableTypesWithoutRecursion? echoNullFilledClass =
+          await api.echoAllNullableTypesWithoutRecursion(nullableListTypes);
+
+      compareAllNullableTypesWithoutRecursion(
+          nullableListTypes, echoNullFilledClass);
+    });
+
     testWidgets('errors are returned correctly', (WidgetTester _) async {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
 
@@ -266,7 +446,10 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
 
       final AllClassesWrapper sentObject = AllClassesWrapper(
-          allNullableTypes: genericAllNullableTypes, allTypes: genericAllTypes);
+          allNullableTypes: recursiveAllNullableTypes,
+          allNullableTypesWithoutRecursion:
+              genericAllNullableTypesWithoutRecursion,
+          allTypes: genericAllTypes);
 
       final String? receivedString =
           await api.extractNestedNullableString(sentObject);
@@ -288,7 +471,10 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
 
       final AllClassesWrapper sentWrapper = AllClassesWrapper(
-          allNullableTypes: AllNullableTypes(), allTypes: genericAllTypes);
+        allNullableTypes: AllNullableTypes(),
+        allNullableTypesWithoutRecursion: AllNullableTypesWithoutRecursion(),
+        allTypes: genericAllTypes,
+      );
 
       final AllClassesWrapper receivedClassWrapper =
           await api.echoClassWrapper(sentWrapper);
@@ -299,8 +485,10 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
         (WidgetTester _) async {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
 
-      final AllClassesWrapper sentWrapper =
-          AllClassesWrapper(allNullableTypes: AllNullableTypes());
+      final AllClassesWrapper sentWrapper = AllClassesWrapper(
+        allNullableTypes: AllNullableTypes(),
+        allNullableTypesWithoutRecursion: AllNullableTypesWithoutRecursion(),
+      );
 
       final AllClassesWrapper receivedClassWrapper =
           await api.echoClassWrapper(sentWrapper);
@@ -329,6 +517,34 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
 
       final AllNullableTypes echoNullFilledClass =
           await api.sendMultipleNullableTypes(null, null, null);
+      expect(echoNullFilledClass.aNullableInt, null);
+      expect(echoNullFilledClass.aNullableBool, null);
+      expect(echoNullFilledClass.aNullableString, null);
+    });
+
+    testWidgets(
+        'Arguments of multiple types serialize and deserialize correctly (WithoutRecursion)',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+      const String aNullableString = 'this is a String';
+      const bool aNullableBool = false;
+      const int aNullableInt = _regularInt;
+
+      final AllNullableTypesWithoutRecursion echoObject =
+          await api.sendMultipleNullableTypesWithoutRecursion(
+              aNullableBool, aNullableInt, aNullableString);
+      expect(echoObject.aNullableInt, aNullableInt);
+      expect(echoObject.aNullableBool, aNullableBool);
+      expect(echoObject.aNullableString, aNullableString);
+    });
+
+    testWidgets(
+        'Arguments of multiple null types serialize and deserialize correctly (WithoutRecursion)',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion echoNullFilledClass =
+          await api.sendMultipleNullableTypesWithoutRecursion(null, null, null);
       expect(echoNullFilledClass.aNullableInt, null);
       expect(echoNullFilledClass.aNullableBool, null);
       expect(echoNullFilledClass.aNullableString, null);
@@ -775,10 +991,10 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
         (WidgetTester _) async {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
 
-      final AllNullableTypes? echoObject =
-          await api.echoAsyncNullableAllNullableTypes(genericAllNullableTypes);
+      final AllNullableTypes? echoObject = await api
+          .echoAsyncNullableAllNullableTypes(recursiveAllNullableTypes);
 
-      compareAllNullableTypes(echoObject, genericAllNullableTypes);
+      compareAllNullableTypes(echoObject, recursiveAllNullableTypes);
     });
 
     testWidgets('all null datatypes async serialize and deserialize correctly',
@@ -790,6 +1006,33 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       final AllNullableTypes? echoNullFilledClass =
           await api.echoAsyncNullableAllNullableTypes(allTypesNull);
       compareAllNullableTypes(echoNullFilledClass, allTypesNull);
+    });
+
+    testWidgets(
+        'all nullable async datatypes without recursion serialize and deserialize correctly',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion? echoObject =
+          await api.echoAsyncNullableAllNullableTypesWithoutRecursion(
+              genericAllNullableTypesWithoutRecursion);
+
+      compareAllNullableTypesWithoutRecursion(
+          echoObject, genericAllNullableTypesWithoutRecursion);
+    });
+
+    testWidgets(
+        'all null datatypes without recursion async serialize and deserialize correctly',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion allTypesNull =
+          AllNullableTypesWithoutRecursion();
+
+      final AllNullableTypesWithoutRecursion? echoNullFilledClass = await api
+          .echoAsyncNullableAllNullableTypesWithoutRecursion(allTypesNull);
+      compareAllNullableTypesWithoutRecursion(
+          echoNullFilledClass, allTypesNull);
     });
 
     testWidgets('Int async serialize and deserialize correctly',
@@ -1187,6 +1430,35 @@ void runPigeonIntegrationTests(TargetGenerator targetGenerator) {
       expect(compositeObject.aNullableString, null);
     });
 
+    testWidgets(
+        'Arguments of multiple types serialize and deserialize correctly (WithoutRecursion)',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+      const String aNullableString = 'this is a String';
+      const bool aNullableBool = false;
+      const int aNullableInt = _regularInt;
+
+      final AllNullableTypesWithoutRecursion compositeObject =
+          await api.callFlutterSendMultipleNullableTypesWithoutRecursion(
+              aNullableBool, aNullableInt, aNullableString);
+      expect(compositeObject.aNullableInt, aNullableInt);
+      expect(compositeObject.aNullableBool, aNullableBool);
+      expect(compositeObject.aNullableString, aNullableString);
+    });
+
+    testWidgets(
+        'Arguments of multiple null types serialize and deserialize correctly (WithoutRecursion)',
+        (WidgetTester _) async {
+      final HostIntegrationCoreApi api = HostIntegrationCoreApi();
+
+      final AllNullableTypesWithoutRecursion compositeObject =
+          await api.callFlutterSendMultipleNullableTypesWithoutRecursion(
+              null, null, null);
+      expect(compositeObject.aNullableInt, null);
+      expect(compositeObject.aNullableBool, null);
+      expect(compositeObject.aNullableString, null);
+    });
+
     testWidgets('booleans serialize and deserialize correctly',
         (WidgetTester _) async {
       final HostIntegrationCoreApi api = HostIntegrationCoreApi();
@@ -1484,6 +1756,12 @@ class _FlutterApiTestImplementation implements FlutterIntegrationCoreApi {
   }
 
   @override
+  AllNullableTypesWithoutRecursion? echoAllNullableTypesWithoutRecursion(
+      AllNullableTypesWithoutRecursion? everything) {
+    return everything;
+  }
+
+  @override
   void noop() {}
 
   @override
@@ -1500,6 +1778,15 @@ class _FlutterApiTestImplementation implements FlutterIntegrationCoreApi {
   AllNullableTypes sendMultipleNullableTypes(
       bool? aNullableBool, int? aNullableInt, String? aNullableString) {
     return AllNullableTypes(
+        aNullableBool: aNullableBool,
+        aNullableInt: aNullableInt,
+        aNullableString: aNullableString);
+  }
+
+  @override
+  AllNullableTypesWithoutRecursion sendMultipleNullableTypesWithoutRecursion(
+      bool? aNullableBool, int? aNullableInt, String? aNullableString) {
+    return AllNullableTypesWithoutRecursion(
         aNullableBool: aNullableBool,
         aNullableInt: aNullableInt,
         aNullableString: aNullableString);
