@@ -4,19 +4,18 @@
 
 package com.example.test_plugin
 
-import androidx.annotation.NonNull
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 
 /** This plugin handles the native side of the integration tests in example/integration_test/. */
 class TestPlugin : FlutterPlugin, HostIntegrationCoreApi {
   var flutterApi: FlutterIntegrationCoreApi? = null
 
-  override fun onAttachedToEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
-    HostIntegrationCoreApi.setUp(binding.getBinaryMessenger(), this)
-    flutterApi = FlutterIntegrationCoreApi(binding.getBinaryMessenger())
+  override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+    HostIntegrationCoreApi.setUp(binding.binaryMessenger, this)
+    flutterApi = FlutterIntegrationCoreApi(binding.binaryMessenger)
   }
 
-  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {}
+  override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {}
 
   // HostIntegrationCoreApi
 
@@ -27,6 +26,12 @@ class TestPlugin : FlutterPlugin, HostIntegrationCoreApi {
   }
 
   override fun echoAllNullableTypes(everything: AllNullableTypes?): AllNullableTypes? {
+    return everything
+  }
+
+  override fun echoAllNullableTypesWithoutRecursion(
+      everything: AllNullableTypesWithoutRecursion?
+  ): AllNullableTypesWithoutRecursion? {
     return everything
   }
 
@@ -113,6 +118,17 @@ class TestPlugin : FlutterPlugin, HostIntegrationCoreApi {
         aNullableString = aNullableString)
   }
 
+  override fun sendMultipleNullableTypesWithoutRecursion(
+      aNullableBool: Boolean?,
+      aNullableInt: Long?,
+      aNullableString: String?
+  ): AllNullableTypesWithoutRecursion {
+    return AllNullableTypesWithoutRecursion(
+        aNullableBool = aNullableBool,
+        aNullableInt = aNullableInt,
+        aNullableString = aNullableString)
+  }
+
   override fun echoNullableInt(aNullableInt: Long?): Long? {
     return aNullableInt
   }
@@ -180,6 +196,13 @@ class TestPlugin : FlutterPlugin, HostIntegrationCoreApi {
   override fun echoAsyncNullableAllNullableTypes(
       everything: AllNullableTypes?,
       callback: (Result<AllNullableTypes?>) -> Unit
+  ) {
+    callback(Result.success(everything))
+  }
+
+  override fun echoAsyncNullableAllNullableTypesWithoutRecursion(
+      everything: AllNullableTypesWithoutRecursion?,
+      callback: (Result<AllNullableTypesWithoutRecursion?>) -> Unit
   ) {
     callback(Result.success(everything))
   }
@@ -290,6 +313,25 @@ class TestPlugin : FlutterPlugin, HostIntegrationCoreApi {
     flutterApi!!.sendMultipleNullableTypes(aNullableBool, aNullableInt, aNullableString) { echo ->
       callback(echo)
     }
+  }
+
+  override fun callFlutterEchoAllNullableTypesWithoutRecursion(
+      everything: AllNullableTypesWithoutRecursion?,
+      callback: (Result<AllNullableTypesWithoutRecursion?>) -> Unit
+  ) {
+    flutterApi!!.echoAllNullableTypesWithoutRecursion(everything) { echo -> callback(echo) }
+  }
+
+  override fun callFlutterSendMultipleNullableTypesWithoutRecursion(
+      aNullableBool: Boolean?,
+      aNullableInt: Long?,
+      aNullableString: String?,
+      callback: (Result<AllNullableTypesWithoutRecursion>) -> Unit
+  ) {
+    flutterApi!!.sendMultipleNullableTypesWithoutRecursion(
+        aNullableBool, aNullableInt, aNullableString) { echo ->
+          callback(echo)
+        }
   }
 
   override fun callFlutterEchoBool(aBool: Boolean, callback: (Result<Boolean>) -> Unit) {
