@@ -177,7 +177,8 @@ AllNullableTypes::AllNullableTypes(
     const EncodableMap* nullable_map_with_annotations,
     const EncodableMap* nullable_map_with_object, const AnEnum* a_nullable_enum,
     const std::string* a_nullable_string,
-    const EncodableValue* a_nullable_object)
+    const EncodableValue* a_nullable_object,
+    const AllNullableTypes* all_nullable_types)
     : a_nullable_bool_(a_nullable_bool ? std::optional<bool>(*a_nullable_bool)
                                        : std::nullopt),
       a_nullable_int_(a_nullable_int ? std::optional<int64_t>(*a_nullable_int)
@@ -228,7 +229,99 @@ AllNullableTypes::AllNullableTypes(
                              : std::nullopt),
       a_nullable_object_(a_nullable_object
                              ? std::optional<EncodableValue>(*a_nullable_object)
-                             : std::nullopt) {}
+                             : std::nullopt),
+      all_nullable_types_(
+          all_nullable_types
+              ? std::make_unique<AllNullableTypes>(*all_nullable_types)
+              : nullptr) {}
+
+AllNullableTypes::AllNullableTypes(const AllNullableTypes& other)
+    : a_nullable_bool_(other.a_nullable_bool_
+                           ? std::optional<bool>(*other.a_nullable_bool_)
+                           : std::nullopt),
+      a_nullable_int_(other.a_nullable_int_
+                          ? std::optional<int64_t>(*other.a_nullable_int_)
+                          : std::nullopt),
+      a_nullable_int64_(other.a_nullable_int64_
+                            ? std::optional<int64_t>(*other.a_nullable_int64_)
+                            : std::nullopt),
+      a_nullable_double_(other.a_nullable_double_
+                             ? std::optional<double>(*other.a_nullable_double_)
+                             : std::nullopt),
+      a_nullable_byte_array_(other.a_nullable_byte_array_
+                                 ? std::optional<std::vector<uint8_t>>(
+                                       *other.a_nullable_byte_array_)
+                                 : std::nullopt),
+      a_nullable4_byte_array_(other.a_nullable4_byte_array_
+                                  ? std::optional<std::vector<int32_t>>(
+                                        *other.a_nullable4_byte_array_)
+                                  : std::nullopt),
+      a_nullable8_byte_array_(other.a_nullable8_byte_array_
+                                  ? std::optional<std::vector<int64_t>>(
+                                        *other.a_nullable8_byte_array_)
+                                  : std::nullopt),
+      a_nullable_float_array_(other.a_nullable_float_array_
+                                  ? std::optional<std::vector<double>>(
+                                        *other.a_nullable_float_array_)
+                                  : std::nullopt),
+      a_nullable_list_(other.a_nullable_list_ ? std::optional<EncodableList>(
+                                                    *other.a_nullable_list_)
+                                              : std::nullopt),
+      a_nullable_map_(other.a_nullable_map_
+                          ? std::optional<EncodableMap>(*other.a_nullable_map_)
+                          : std::nullopt),
+      nullable_nested_list_(
+          other.nullable_nested_list_
+              ? std::optional<EncodableList>(*other.nullable_nested_list_)
+              : std::nullopt),
+      nullable_map_with_annotations_(
+          other.nullable_map_with_annotations_
+              ? std::optional<EncodableMap>(
+                    *other.nullable_map_with_annotations_)
+              : std::nullopt),
+      nullable_map_with_object_(
+          other.nullable_map_with_object_
+              ? std::optional<EncodableMap>(*other.nullable_map_with_object_)
+              : std::nullopt),
+      a_nullable_enum_(other.a_nullable_enum_
+                           ? std::optional<AnEnum>(*other.a_nullable_enum_)
+                           : std::nullopt),
+      a_nullable_string_(
+          other.a_nullable_string_
+              ? std::optional<std::string>(*other.a_nullable_string_)
+              : std::nullopt),
+      a_nullable_object_(
+          other.a_nullable_object_
+              ? std::optional<EncodableValue>(*other.a_nullable_object_)
+              : std::nullopt),
+      all_nullable_types_(
+          other.all_nullable_types_
+              ? std::make_unique<AllNullableTypes>(*other.all_nullable_types_)
+              : nullptr) {}
+
+AllNullableTypes& AllNullableTypes::operator=(const AllNullableTypes& other) {
+  a_nullable_bool_ = other.a_nullable_bool_;
+  a_nullable_int_ = other.a_nullable_int_;
+  a_nullable_int64_ = other.a_nullable_int64_;
+  a_nullable_double_ = other.a_nullable_double_;
+  a_nullable_byte_array_ = other.a_nullable_byte_array_;
+  a_nullable4_byte_array_ = other.a_nullable4_byte_array_;
+  a_nullable8_byte_array_ = other.a_nullable8_byte_array_;
+  a_nullable_float_array_ = other.a_nullable_float_array_;
+  a_nullable_list_ = other.a_nullable_list_;
+  a_nullable_map_ = other.a_nullable_map_;
+  nullable_nested_list_ = other.nullable_nested_list_;
+  nullable_map_with_annotations_ = other.nullable_map_with_annotations_;
+  nullable_map_with_object_ = other.nullable_map_with_object_;
+  a_nullable_enum_ = other.a_nullable_enum_;
+  a_nullable_string_ = other.a_nullable_string_;
+  a_nullable_object_ = other.a_nullable_object_;
+  all_nullable_types_ =
+      other.all_nullable_types_
+          ? std::make_unique<AllNullableTypes>(*other.all_nullable_types_)
+          : nullptr;
+  return *this;
+}
 
 const bool* AllNullableTypes::a_nullable_bool() const {
   return a_nullable_bool_ ? &(*a_nullable_bool_) : nullptr;
@@ -456,9 +549,24 @@ void AllNullableTypes::set_a_nullable_object(const EncodableValue& value_arg) {
   a_nullable_object_ = value_arg;
 }
 
+const AllNullableTypes* AllNullableTypes::all_nullable_types() const {
+  return all_nullable_types_.get();
+}
+
+void AllNullableTypes::set_all_nullable_types(
+    const AllNullableTypes* value_arg) {
+  all_nullable_types_ =
+      value_arg ? std::make_unique<AllNullableTypes>(*value_arg) : nullptr;
+}
+
+void AllNullableTypes::set_all_nullable_types(
+    const AllNullableTypes& value_arg) {
+  all_nullable_types_ = std::make_unique<AllNullableTypes>(value_arg);
+}
+
 EncodableList AllNullableTypes::ToEncodableList() const {
   EncodableList list;
-  list.reserve(16);
+  list.reserve(17);
   list.push_back(a_nullable_bool_ ? EncodableValue(*a_nullable_bool_)
                                   : EncodableValue());
   list.push_back(a_nullable_int_ ? EncodableValue(*a_nullable_int_)
@@ -496,6 +604,9 @@ EncodableList AllNullableTypes::ToEncodableList() const {
   list.push_back(a_nullable_string_ ? EncodableValue(*a_nullable_string_)
                                     : EncodableValue());
   list.push_back(a_nullable_object_ ? *a_nullable_object_ : EncodableValue());
+  list.push_back(all_nullable_types_
+                     ? EncodableValue(all_nullable_types_->ToEncodableList())
+                     : EncodableValue());
   return list;
 }
 
@@ -578,45 +689,549 @@ AllNullableTypes AllNullableTypes::FromEncodableList(
   if (!encodable_a_nullable_object.IsNull()) {
     decoded.set_a_nullable_object(encodable_a_nullable_object);
   }
+  auto& encodable_all_nullable_types = list[16];
+  if (!encodable_all_nullable_types.IsNull()) {
+    decoded.set_all_nullable_types(AllNullableTypes::FromEncodableList(
+        std::get<EncodableList>(encodable_all_nullable_types)));
+  }
+  return decoded;
+}
+
+// AllNullableTypesWithoutRecursion
+
+AllNullableTypesWithoutRecursion::AllNullableTypesWithoutRecursion() {}
+
+AllNullableTypesWithoutRecursion::AllNullableTypesWithoutRecursion(
+    const bool* a_nullable_bool, const int64_t* a_nullable_int,
+    const int64_t* a_nullable_int64, const double* a_nullable_double,
+    const std::vector<uint8_t>* a_nullable_byte_array,
+    const std::vector<int32_t>* a_nullable4_byte_array,
+    const std::vector<int64_t>* a_nullable8_byte_array,
+    const std::vector<double>* a_nullable_float_array,
+    const EncodableList* a_nullable_list, const EncodableMap* a_nullable_map,
+    const EncodableList* nullable_nested_list,
+    const EncodableMap* nullable_map_with_annotations,
+    const EncodableMap* nullable_map_with_object, const AnEnum* a_nullable_enum,
+    const std::string* a_nullable_string,
+    const EncodableValue* a_nullable_object)
+    : a_nullable_bool_(a_nullable_bool ? std::optional<bool>(*a_nullable_bool)
+                                       : std::nullopt),
+      a_nullable_int_(a_nullable_int ? std::optional<int64_t>(*a_nullable_int)
+                                     : std::nullopt),
+      a_nullable_int64_(a_nullable_int64
+                            ? std::optional<int64_t>(*a_nullable_int64)
+                            : std::nullopt),
+      a_nullable_double_(a_nullable_double
+                             ? std::optional<double>(*a_nullable_double)
+                             : std::nullopt),
+      a_nullable_byte_array_(
+          a_nullable_byte_array
+              ? std::optional<std::vector<uint8_t>>(*a_nullable_byte_array)
+              : std::nullopt),
+      a_nullable4_byte_array_(
+          a_nullable4_byte_array
+              ? std::optional<std::vector<int32_t>>(*a_nullable4_byte_array)
+              : std::nullopt),
+      a_nullable8_byte_array_(
+          a_nullable8_byte_array
+              ? std::optional<std::vector<int64_t>>(*a_nullable8_byte_array)
+              : std::nullopt),
+      a_nullable_float_array_(
+          a_nullable_float_array
+              ? std::optional<std::vector<double>>(*a_nullable_float_array)
+              : std::nullopt),
+      a_nullable_list_(a_nullable_list
+                           ? std::optional<EncodableList>(*a_nullable_list)
+                           : std::nullopt),
+      a_nullable_map_(a_nullable_map
+                          ? std::optional<EncodableMap>(*a_nullable_map)
+                          : std::nullopt),
+      nullable_nested_list_(nullable_nested_list ? std::optional<EncodableList>(
+                                                       *nullable_nested_list)
+                                                 : std::nullopt),
+      nullable_map_with_annotations_(
+          nullable_map_with_annotations
+              ? std::optional<EncodableMap>(*nullable_map_with_annotations)
+              : std::nullopt),
+      nullable_map_with_object_(
+          nullable_map_with_object
+              ? std::optional<EncodableMap>(*nullable_map_with_object)
+              : std::nullopt),
+      a_nullable_enum_(a_nullable_enum ? std::optional<AnEnum>(*a_nullable_enum)
+                                       : std::nullopt),
+      a_nullable_string_(a_nullable_string
+                             ? std::optional<std::string>(*a_nullable_string)
+                             : std::nullopt),
+      a_nullable_object_(a_nullable_object
+                             ? std::optional<EncodableValue>(*a_nullable_object)
+                             : std::nullopt) {}
+
+const bool* AllNullableTypesWithoutRecursion::a_nullable_bool() const {
+  return a_nullable_bool_ ? &(*a_nullable_bool_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_bool(
+    const bool* value_arg) {
+  a_nullable_bool_ = value_arg ? std::optional<bool>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_bool(bool value_arg) {
+  a_nullable_bool_ = value_arg;
+}
+
+const int64_t* AllNullableTypesWithoutRecursion::a_nullable_int() const {
+  return a_nullable_int_ ? &(*a_nullable_int_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_int(
+    const int64_t* value_arg) {
+  a_nullable_int_ =
+      value_arg ? std::optional<int64_t>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_int(int64_t value_arg) {
+  a_nullable_int_ = value_arg;
+}
+
+const int64_t* AllNullableTypesWithoutRecursion::a_nullable_int64() const {
+  return a_nullable_int64_ ? &(*a_nullable_int64_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_int64(
+    const int64_t* value_arg) {
+  a_nullable_int64_ =
+      value_arg ? std::optional<int64_t>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_int64(int64_t value_arg) {
+  a_nullable_int64_ = value_arg;
+}
+
+const double* AllNullableTypesWithoutRecursion::a_nullable_double() const {
+  return a_nullable_double_ ? &(*a_nullable_double_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_double(
+    const double* value_arg) {
+  a_nullable_double_ =
+      value_arg ? std::optional<double>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_double(double value_arg) {
+  a_nullable_double_ = value_arg;
+}
+
+const std::vector<uint8_t>*
+AllNullableTypesWithoutRecursion::a_nullable_byte_array() const {
+  return a_nullable_byte_array_ ? &(*a_nullable_byte_array_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_byte_array(
+    const std::vector<uint8_t>* value_arg) {
+  a_nullable_byte_array_ = value_arg
+                               ? std::optional<std::vector<uint8_t>>(*value_arg)
+                               : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_byte_array(
+    const std::vector<uint8_t>& value_arg) {
+  a_nullable_byte_array_ = value_arg;
+}
+
+const std::vector<int32_t>*
+AllNullableTypesWithoutRecursion::a_nullable4_byte_array() const {
+  return a_nullable4_byte_array_ ? &(*a_nullable4_byte_array_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable4_byte_array(
+    const std::vector<int32_t>* value_arg) {
+  a_nullable4_byte_array_ =
+      value_arg ? std::optional<std::vector<int32_t>>(*value_arg)
+                : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable4_byte_array(
+    const std::vector<int32_t>& value_arg) {
+  a_nullable4_byte_array_ = value_arg;
+}
+
+const std::vector<int64_t>*
+AllNullableTypesWithoutRecursion::a_nullable8_byte_array() const {
+  return a_nullable8_byte_array_ ? &(*a_nullable8_byte_array_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable8_byte_array(
+    const std::vector<int64_t>* value_arg) {
+  a_nullable8_byte_array_ =
+      value_arg ? std::optional<std::vector<int64_t>>(*value_arg)
+                : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable8_byte_array(
+    const std::vector<int64_t>& value_arg) {
+  a_nullable8_byte_array_ = value_arg;
+}
+
+const std::vector<double>*
+AllNullableTypesWithoutRecursion::a_nullable_float_array() const {
+  return a_nullable_float_array_ ? &(*a_nullable_float_array_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_float_array(
+    const std::vector<double>* value_arg) {
+  a_nullable_float_array_ =
+      value_arg ? std::optional<std::vector<double>>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_float_array(
+    const std::vector<double>& value_arg) {
+  a_nullable_float_array_ = value_arg;
+}
+
+const EncodableList* AllNullableTypesWithoutRecursion::a_nullable_list() const {
+  return a_nullable_list_ ? &(*a_nullable_list_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_list(
+    const EncodableList* value_arg) {
+  a_nullable_list_ =
+      value_arg ? std::optional<EncodableList>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_list(
+    const EncodableList& value_arg) {
+  a_nullable_list_ = value_arg;
+}
+
+const EncodableMap* AllNullableTypesWithoutRecursion::a_nullable_map() const {
+  return a_nullable_map_ ? &(*a_nullable_map_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_map(
+    const EncodableMap* value_arg) {
+  a_nullable_map_ =
+      value_arg ? std::optional<EncodableMap>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_map(
+    const EncodableMap& value_arg) {
+  a_nullable_map_ = value_arg;
+}
+
+const EncodableList* AllNullableTypesWithoutRecursion::nullable_nested_list()
+    const {
+  return nullable_nested_list_ ? &(*nullable_nested_list_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_nullable_nested_list(
+    const EncodableList* value_arg) {
+  nullable_nested_list_ =
+      value_arg ? std::optional<EncodableList>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_nullable_nested_list(
+    const EncodableList& value_arg) {
+  nullable_nested_list_ = value_arg;
+}
+
+const EncodableMap*
+AllNullableTypesWithoutRecursion::nullable_map_with_annotations() const {
+  return nullable_map_with_annotations_ ? &(*nullable_map_with_annotations_)
+                                        : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_nullable_map_with_annotations(
+    const EncodableMap* value_arg) {
+  nullable_map_with_annotations_ =
+      value_arg ? std::optional<EncodableMap>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_nullable_map_with_annotations(
+    const EncodableMap& value_arg) {
+  nullable_map_with_annotations_ = value_arg;
+}
+
+const EncodableMap* AllNullableTypesWithoutRecursion::nullable_map_with_object()
+    const {
+  return nullable_map_with_object_ ? &(*nullable_map_with_object_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_nullable_map_with_object(
+    const EncodableMap* value_arg) {
+  nullable_map_with_object_ =
+      value_arg ? std::optional<EncodableMap>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_nullable_map_with_object(
+    const EncodableMap& value_arg) {
+  nullable_map_with_object_ = value_arg;
+}
+
+const AnEnum* AllNullableTypesWithoutRecursion::a_nullable_enum() const {
+  return a_nullable_enum_ ? &(*a_nullable_enum_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_enum(
+    const AnEnum* value_arg) {
+  a_nullable_enum_ =
+      value_arg ? std::optional<AnEnum>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_enum(
+    const AnEnum& value_arg) {
+  a_nullable_enum_ = value_arg;
+}
+
+const std::string* AllNullableTypesWithoutRecursion::a_nullable_string() const {
+  return a_nullable_string_ ? &(*a_nullable_string_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_string(
+    const std::string_view* value_arg) {
+  a_nullable_string_ =
+      value_arg ? std::optional<std::string>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_string(
+    std::string_view value_arg) {
+  a_nullable_string_ = value_arg;
+}
+
+const EncodableValue* AllNullableTypesWithoutRecursion::a_nullable_object()
+    const {
+  return a_nullable_object_ ? &(*a_nullable_object_) : nullptr;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_object(
+    const EncodableValue* value_arg) {
+  a_nullable_object_ =
+      value_arg ? std::optional<EncodableValue>(*value_arg) : std::nullopt;
+}
+
+void AllNullableTypesWithoutRecursion::set_a_nullable_object(
+    const EncodableValue& value_arg) {
+  a_nullable_object_ = value_arg;
+}
+
+EncodableList AllNullableTypesWithoutRecursion::ToEncodableList() const {
+  EncodableList list;
+  list.reserve(16);
+  list.push_back(a_nullable_bool_ ? EncodableValue(*a_nullable_bool_)
+                                  : EncodableValue());
+  list.push_back(a_nullable_int_ ? EncodableValue(*a_nullable_int_)
+                                 : EncodableValue());
+  list.push_back(a_nullable_int64_ ? EncodableValue(*a_nullable_int64_)
+                                   : EncodableValue());
+  list.push_back(a_nullable_double_ ? EncodableValue(*a_nullable_double_)
+                                    : EncodableValue());
+  list.push_back(a_nullable_byte_array_
+                     ? EncodableValue(*a_nullable_byte_array_)
+                     : EncodableValue());
+  list.push_back(a_nullable4_byte_array_
+                     ? EncodableValue(*a_nullable4_byte_array_)
+                     : EncodableValue());
+  list.push_back(a_nullable8_byte_array_
+                     ? EncodableValue(*a_nullable8_byte_array_)
+                     : EncodableValue());
+  list.push_back(a_nullable_float_array_
+                     ? EncodableValue(*a_nullable_float_array_)
+                     : EncodableValue());
+  list.push_back(a_nullable_list_ ? EncodableValue(*a_nullable_list_)
+                                  : EncodableValue());
+  list.push_back(a_nullable_map_ ? EncodableValue(*a_nullable_map_)
+                                 : EncodableValue());
+  list.push_back(nullable_nested_list_ ? EncodableValue(*nullable_nested_list_)
+                                       : EncodableValue());
+  list.push_back(nullable_map_with_annotations_
+                     ? EncodableValue(*nullable_map_with_annotations_)
+                     : EncodableValue());
+  list.push_back(nullable_map_with_object_
+                     ? EncodableValue(*nullable_map_with_object_)
+                     : EncodableValue());
+  list.push_back(a_nullable_enum_ ? EncodableValue((int)(*a_nullable_enum_))
+                                  : EncodableValue());
+  list.push_back(a_nullable_string_ ? EncodableValue(*a_nullable_string_)
+                                    : EncodableValue());
+  list.push_back(a_nullable_object_ ? *a_nullable_object_ : EncodableValue());
+  return list;
+}
+
+AllNullableTypesWithoutRecursion
+AllNullableTypesWithoutRecursion::FromEncodableList(const EncodableList& list) {
+  AllNullableTypesWithoutRecursion decoded;
+  auto& encodable_a_nullable_bool = list[0];
+  if (!encodable_a_nullable_bool.IsNull()) {
+    decoded.set_a_nullable_bool(std::get<bool>(encodable_a_nullable_bool));
+  }
+  auto& encodable_a_nullable_int = list[1];
+  if (!encodable_a_nullable_int.IsNull()) {
+    decoded.set_a_nullable_int(encodable_a_nullable_int.LongValue());
+  }
+  auto& encodable_a_nullable_int64 = list[2];
+  if (!encodable_a_nullable_int64.IsNull()) {
+    decoded.set_a_nullable_int64(encodable_a_nullable_int64.LongValue());
+  }
+  auto& encodable_a_nullable_double = list[3];
+  if (!encodable_a_nullable_double.IsNull()) {
+    decoded.set_a_nullable_double(
+        std::get<double>(encodable_a_nullable_double));
+  }
+  auto& encodable_a_nullable_byte_array = list[4];
+  if (!encodable_a_nullable_byte_array.IsNull()) {
+    decoded.set_a_nullable_byte_array(
+        std::get<std::vector<uint8_t>>(encodable_a_nullable_byte_array));
+  }
+  auto& encodable_a_nullable4_byte_array = list[5];
+  if (!encodable_a_nullable4_byte_array.IsNull()) {
+    decoded.set_a_nullable4_byte_array(
+        std::get<std::vector<int32_t>>(encodable_a_nullable4_byte_array));
+  }
+  auto& encodable_a_nullable8_byte_array = list[6];
+  if (!encodable_a_nullable8_byte_array.IsNull()) {
+    decoded.set_a_nullable8_byte_array(
+        std::get<std::vector<int64_t>>(encodable_a_nullable8_byte_array));
+  }
+  auto& encodable_a_nullable_float_array = list[7];
+  if (!encodable_a_nullable_float_array.IsNull()) {
+    decoded.set_a_nullable_float_array(
+        std::get<std::vector<double>>(encodable_a_nullable_float_array));
+  }
+  auto& encodable_a_nullable_list = list[8];
+  if (!encodable_a_nullable_list.IsNull()) {
+    decoded.set_a_nullable_list(
+        std::get<EncodableList>(encodable_a_nullable_list));
+  }
+  auto& encodable_a_nullable_map = list[9];
+  if (!encodable_a_nullable_map.IsNull()) {
+    decoded.set_a_nullable_map(
+        std::get<EncodableMap>(encodable_a_nullable_map));
+  }
+  auto& encodable_nullable_nested_list = list[10];
+  if (!encodable_nullable_nested_list.IsNull()) {
+    decoded.set_nullable_nested_list(
+        std::get<EncodableList>(encodable_nullable_nested_list));
+  }
+  auto& encodable_nullable_map_with_annotations = list[11];
+  if (!encodable_nullable_map_with_annotations.IsNull()) {
+    decoded.set_nullable_map_with_annotations(
+        std::get<EncodableMap>(encodable_nullable_map_with_annotations));
+  }
+  auto& encodable_nullable_map_with_object = list[12];
+  if (!encodable_nullable_map_with_object.IsNull()) {
+    decoded.set_nullable_map_with_object(
+        std::get<EncodableMap>(encodable_nullable_map_with_object));
+  }
+  auto& encodable_a_nullable_enum = list[13];
+  if (!encodable_a_nullable_enum.IsNull()) {
+    decoded.set_a_nullable_enum(
+        (AnEnum)(std::get<int32_t>(encodable_a_nullable_enum)));
+  }
+  auto& encodable_a_nullable_string = list[14];
+  if (!encodable_a_nullable_string.IsNull()) {
+    decoded.set_a_nullable_string(
+        std::get<std::string>(encodable_a_nullable_string));
+  }
+  auto& encodable_a_nullable_object = list[15];
+  if (!encodable_a_nullable_object.IsNull()) {
+    decoded.set_a_nullable_object(encodable_a_nullable_object);
+  }
   return decoded;
 }
 
 // AllClassesWrapper
 
 AllClassesWrapper::AllClassesWrapper(const AllNullableTypes& all_nullable_types)
-    : all_nullable_types_(all_nullable_types) {}
+    : all_nullable_types_(
+          std::make_unique<AllNullableTypes>(all_nullable_types)) {}
 
 AllClassesWrapper::AllClassesWrapper(const AllNullableTypes& all_nullable_types,
+                                     const AllNullableTypesWithoutRecursion*
+                                         all_nullable_types_without_recursion,
                                      const AllTypes* all_types)
-    : all_nullable_types_(all_nullable_types),
-      all_types_(all_types ? std::optional<AllTypes>(*all_types)
-                           : std::nullopt) {}
+    : all_nullable_types_(
+          std::make_unique<AllNullableTypes>(all_nullable_types)),
+      all_nullable_types_without_recursion_(
+          all_nullable_types_without_recursion
+              ? std::make_unique<AllNullableTypesWithoutRecursion>(
+                    *all_nullable_types_without_recursion)
+              : nullptr),
+      all_types_(all_types ? std::make_unique<AllTypes>(*all_types) : nullptr) {
+}
+
+AllClassesWrapper::AllClassesWrapper(const AllClassesWrapper& other)
+    : all_nullable_types_(
+          std::make_unique<AllNullableTypes>(*other.all_nullable_types_)),
+      all_nullable_types_without_recursion_(
+          other.all_nullable_types_without_recursion_
+              ? std::make_unique<AllNullableTypesWithoutRecursion>(
+                    *other.all_nullable_types_without_recursion_)
+              : nullptr),
+      all_types_(other.all_types_
+                     ? std::make_unique<AllTypes>(*other.all_types_)
+                     : nullptr) {}
+
+AllClassesWrapper& AllClassesWrapper::operator=(
+    const AllClassesWrapper& other) {
+  all_nullable_types_ =
+      std::make_unique<AllNullableTypes>(*other.all_nullable_types_);
+  all_nullable_types_without_recursion_ =
+      other.all_nullable_types_without_recursion_
+          ? std::make_unique<AllNullableTypesWithoutRecursion>(
+                *other.all_nullable_types_without_recursion_)
+          : nullptr;
+  all_types_ = other.all_types_ ? std::make_unique<AllTypes>(*other.all_types_)
+                                : nullptr;
+  return *this;
+}
 
 const AllNullableTypes& AllClassesWrapper::all_nullable_types() const {
-  return all_nullable_types_;
+  return *all_nullable_types_;
 }
 
 void AllClassesWrapper::set_all_nullable_types(
     const AllNullableTypes& value_arg) {
-  all_nullable_types_ = value_arg;
+  all_nullable_types_ = std::make_unique<AllNullableTypes>(value_arg);
+}
+
+const AllNullableTypesWithoutRecursion*
+AllClassesWrapper::all_nullable_types_without_recursion() const {
+  return all_nullable_types_without_recursion_.get();
+}
+
+void AllClassesWrapper::set_all_nullable_types_without_recursion(
+    const AllNullableTypesWithoutRecursion* value_arg) {
+  all_nullable_types_without_recursion_ =
+      value_arg ? std::make_unique<AllNullableTypesWithoutRecursion>(*value_arg)
+                : nullptr;
+}
+
+void AllClassesWrapper::set_all_nullable_types_without_recursion(
+    const AllNullableTypesWithoutRecursion& value_arg) {
+  all_nullable_types_without_recursion_ =
+      std::make_unique<AllNullableTypesWithoutRecursion>(value_arg);
 }
 
 const AllTypes* AllClassesWrapper::all_types() const {
-  return all_types_ ? &(*all_types_) : nullptr;
+  return all_types_.get();
 }
 
 void AllClassesWrapper::set_all_types(const AllTypes* value_arg) {
-  all_types_ = value_arg ? std::optional<AllTypes>(*value_arg) : std::nullopt;
+  all_types_ = value_arg ? std::make_unique<AllTypes>(*value_arg) : nullptr;
 }
 
 void AllClassesWrapper::set_all_types(const AllTypes& value_arg) {
-  all_types_ = value_arg;
+  all_types_ = std::make_unique<AllTypes>(value_arg);
 }
 
 EncodableList AllClassesWrapper::ToEncodableList() const {
   EncodableList list;
-  list.reserve(2);
-  list.push_back(EncodableValue(all_nullable_types_.ToEncodableList()));
+  list.reserve(3);
+  list.push_back(EncodableValue(all_nullable_types_->ToEncodableList()));
+  list.push_back(
+      all_nullable_types_without_recursion_
+          ? EncodableValue(
+                all_nullable_types_without_recursion_->ToEncodableList())
+          : EncodableValue());
   list.push_back(all_types_ ? EncodableValue(all_types_->ToEncodableList())
                             : EncodableValue());
   return list;
@@ -626,7 +1241,14 @@ AllClassesWrapper AllClassesWrapper::FromEncodableList(
     const EncodableList& list) {
   AllClassesWrapper decoded(
       AllNullableTypes::FromEncodableList(std::get<EncodableList>(list[0])));
-  auto& encodable_all_types = list[1];
+  auto& encodable_all_nullable_types_without_recursion = list[1];
+  if (!encodable_all_nullable_types_without_recursion.IsNull()) {
+    decoded.set_all_nullable_types_without_recursion(
+        AllNullableTypesWithoutRecursion::FromEncodableList(
+            std::get<EncodableList>(
+                encodable_all_nullable_types_without_recursion)));
+  }
+  auto& encodable_all_types = list[2];
   if (!encodable_all_types.IsNull()) {
     decoded.set_all_types(AllTypes::FromEncodableList(
         std::get<EncodableList>(encodable_all_types)));
@@ -684,9 +1306,13 @@ EncodableValue HostIntegrationCoreApiCodecSerializer::ReadValueOfType(
       return CustomEncodableValue(AllNullableTypes::FromEncodableList(
           std::get<EncodableList>(ReadValue(stream))));
     case 130:
+      return CustomEncodableValue(
+          AllNullableTypesWithoutRecursion::FromEncodableList(
+              std::get<EncodableList>(ReadValue(stream))));
+    case 131:
       return CustomEncodableValue(AllTypes::FromEncodableList(
           std::get<EncodableList>(ReadValue(stream))));
-    case 131:
+    case 132:
       return CustomEncodableValue(TestMessage::FromEncodableList(
           std::get<EncodableList>(ReadValue(stream))));
     default:
@@ -713,15 +1339,23 @@ void HostIntegrationCoreApiCodecSerializer::WriteValue(
           stream);
       return;
     }
-    if (custom_value->type() == typeid(AllTypes)) {
+    if (custom_value->type() == typeid(AllNullableTypesWithoutRecursion)) {
       stream->WriteByte(130);
+      WriteValue(EncodableValue(std::any_cast<AllNullableTypesWithoutRecursion>(
+                                    *custom_value)
+                                    .ToEncodableList()),
+                 stream);
+      return;
+    }
+    if (custom_value->type() == typeid(AllTypes)) {
+      stream->WriteByte(131);
       WriteValue(EncodableValue(
                      std::any_cast<AllTypes>(*custom_value).ToEncodableList()),
                  stream);
       return;
     }
     if (custom_value->type() == typeid(TestMessage)) {
-      stream->WriteByte(131);
+      stream->WriteByte(132);
       WriteValue(
           EncodableValue(
               std::any_cast<TestMessage>(*custom_value).ToEncodableList()),
@@ -1383,6 +2017,46 @@ void HostIntegrationCoreApi::SetUp(flutter::BinaryMessenger* binary_messenger,
     BasicMessageChannel<> channel(
         binary_messenger,
         "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi."
+        "echoAllNullableTypesWithoutRecursion",
+        &GetCodec());
+    if (api != nullptr) {
+      channel.SetMessageHandler(
+          [api](const EncodableValue& message,
+                const flutter::MessageReply<EncodableValue>& reply) {
+            try {
+              const auto& args = std::get<EncodableList>(message);
+              const auto& encodable_everything_arg = args.at(0);
+              const auto* everything_arg =
+                  &(std::any_cast<const AllNullableTypesWithoutRecursion&>(
+                      std::get<CustomEncodableValue>(
+                          encodable_everything_arg)));
+              ErrorOr<std::optional<AllNullableTypesWithoutRecursion>> output =
+                  api->EchoAllNullableTypesWithoutRecursion(everything_arg);
+              if (output.has_error()) {
+                reply(WrapError(output.error()));
+                return;
+              }
+              EncodableList wrapped;
+              auto output_optional = std::move(output).TakeValue();
+              if (output_optional) {
+                wrapped.push_back(
+                    CustomEncodableValue(std::move(output_optional).value()));
+              } else {
+                wrapped.push_back(EncodableValue());
+              }
+              reply(EncodableValue(std::move(wrapped)));
+            } catch (const std::exception& exception) {
+              reply(WrapError(exception.what()));
+            }
+          });
+    } else {
+      channel.SetMessageHandler(nullptr);
+    }
+  }
+  {
+    BasicMessageChannel<> channel(
+        binary_messenger,
+        "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi."
         "extractNestedNullableString",
         &GetCodec());
     if (api != nullptr) {
@@ -1484,6 +2158,53 @@ void HostIntegrationCoreApi::SetUp(flutter::BinaryMessenger* binary_messenger,
               ErrorOr<AllNullableTypes> output = api->SendMultipleNullableTypes(
                   a_nullable_bool_arg, a_nullable_int_arg,
                   a_nullable_string_arg);
+              if (output.has_error()) {
+                reply(WrapError(output.error()));
+                return;
+              }
+              EncodableList wrapped;
+              wrapped.push_back(
+                  CustomEncodableValue(std::move(output).TakeValue()));
+              reply(EncodableValue(std::move(wrapped)));
+            } catch (const std::exception& exception) {
+              reply(WrapError(exception.what()));
+            }
+          });
+    } else {
+      channel.SetMessageHandler(nullptr);
+    }
+  }
+  {
+    BasicMessageChannel<> channel(
+        binary_messenger,
+        "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi."
+        "sendMultipleNullableTypesWithoutRecursion",
+        &GetCodec());
+    if (api != nullptr) {
+      channel.SetMessageHandler(
+          [api](const EncodableValue& message,
+                const flutter::MessageReply<EncodableValue>& reply) {
+            try {
+              const auto& args = std::get<EncodableList>(message);
+              const auto& encodable_a_nullable_bool_arg = args.at(0);
+              const auto* a_nullable_bool_arg =
+                  std::get_if<bool>(&encodable_a_nullable_bool_arg);
+              const auto& encodable_a_nullable_int_arg = args.at(1);
+              const int64_t a_nullable_int_arg_value =
+                  encodable_a_nullable_int_arg.IsNull()
+                      ? 0
+                      : encodable_a_nullable_int_arg.LongValue();
+              const auto* a_nullable_int_arg =
+                  encodable_a_nullable_int_arg.IsNull()
+                      ? nullptr
+                      : &a_nullable_int_arg_value;
+              const auto& encodable_a_nullable_string_arg = args.at(2);
+              const auto* a_nullable_string_arg =
+                  std::get_if<std::string>(&encodable_a_nullable_string_arg);
+              ErrorOr<AllNullableTypesWithoutRecursion> output =
+                  api->SendMultipleNullableTypesWithoutRecursion(
+                      a_nullable_bool_arg, a_nullable_int_arg,
+                      a_nullable_string_arg);
               if (output.has_error()) {
                 reply(WrapError(output.error()));
                 return;
@@ -2464,6 +3185,50 @@ void HostIntegrationCoreApi::SetUp(flutter::BinaryMessenger* binary_messenger,
     }
   }
   {
+    BasicMessageChannel<> channel(
+        binary_messenger,
+        "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi."
+        "echoAsyncNullableAllNullableTypesWithoutRecursion",
+        &GetCodec());
+    if (api != nullptr) {
+      channel.SetMessageHandler(
+          [api](const EncodableValue& message,
+                const flutter::MessageReply<EncodableValue>& reply) {
+            try {
+              const auto& args = std::get<EncodableList>(message);
+              const auto& encodable_everything_arg = args.at(0);
+              const auto* everything_arg =
+                  &(std::any_cast<const AllNullableTypesWithoutRecursion&>(
+                      std::get<CustomEncodableValue>(
+                          encodable_everything_arg)));
+              api->EchoAsyncNullableAllNullableTypesWithoutRecursion(
+                  everything_arg,
+                  [reply](
+                      ErrorOr<std::optional<AllNullableTypesWithoutRecursion>>&&
+                          output) {
+                    if (output.has_error()) {
+                      reply(WrapError(output.error()));
+                      return;
+                    }
+                    EncodableList wrapped;
+                    auto output_optional = std::move(output).TakeValue();
+                    if (output_optional) {
+                      wrapped.push_back(CustomEncodableValue(
+                          std::move(output_optional).value()));
+                    } else {
+                      wrapped.push_back(EncodableValue());
+                    }
+                    reply(EncodableValue(std::move(wrapped)));
+                  });
+            } catch (const std::exception& exception) {
+              reply(WrapError(exception.what()));
+            }
+          });
+    } else {
+      channel.SetMessageHandler(nullptr);
+    }
+  }
+  {
     BasicMessageChannel<> channel(binary_messenger,
                                   "dev.flutter.pigeon.pigeon_integration_tests."
                                   "HostIntegrationCoreApi.echoAsyncNullableInt",
@@ -3033,6 +3798,98 @@ void HostIntegrationCoreApi::SetUp(flutter::BinaryMessenger* binary_messenger,
                   a_nullable_bool_arg, a_nullable_int_arg,
                   a_nullable_string_arg,
                   [reply](ErrorOr<AllNullableTypes>&& output) {
+                    if (output.has_error()) {
+                      reply(WrapError(output.error()));
+                      return;
+                    }
+                    EncodableList wrapped;
+                    wrapped.push_back(
+                        CustomEncodableValue(std::move(output).TakeValue()));
+                    reply(EncodableValue(std::move(wrapped)));
+                  });
+            } catch (const std::exception& exception) {
+              reply(WrapError(exception.what()));
+            }
+          });
+    } else {
+      channel.SetMessageHandler(nullptr);
+    }
+  }
+  {
+    BasicMessageChannel<> channel(
+        binary_messenger,
+        "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi."
+        "callFlutterEchoAllNullableTypesWithoutRecursion",
+        &GetCodec());
+    if (api != nullptr) {
+      channel.SetMessageHandler(
+          [api](const EncodableValue& message,
+                const flutter::MessageReply<EncodableValue>& reply) {
+            try {
+              const auto& args = std::get<EncodableList>(message);
+              const auto& encodable_everything_arg = args.at(0);
+              const auto* everything_arg =
+                  &(std::any_cast<const AllNullableTypesWithoutRecursion&>(
+                      std::get<CustomEncodableValue>(
+                          encodable_everything_arg)));
+              api->CallFlutterEchoAllNullableTypesWithoutRecursion(
+                  everything_arg,
+                  [reply](
+                      ErrorOr<std::optional<AllNullableTypesWithoutRecursion>>&&
+                          output) {
+                    if (output.has_error()) {
+                      reply(WrapError(output.error()));
+                      return;
+                    }
+                    EncodableList wrapped;
+                    auto output_optional = std::move(output).TakeValue();
+                    if (output_optional) {
+                      wrapped.push_back(CustomEncodableValue(
+                          std::move(output_optional).value()));
+                    } else {
+                      wrapped.push_back(EncodableValue());
+                    }
+                    reply(EncodableValue(std::move(wrapped)));
+                  });
+            } catch (const std::exception& exception) {
+              reply(WrapError(exception.what()));
+            }
+          });
+    } else {
+      channel.SetMessageHandler(nullptr);
+    }
+  }
+  {
+    BasicMessageChannel<> channel(
+        binary_messenger,
+        "dev.flutter.pigeon.pigeon_integration_tests.HostIntegrationCoreApi."
+        "callFlutterSendMultipleNullableTypesWithoutRecursion",
+        &GetCodec());
+    if (api != nullptr) {
+      channel.SetMessageHandler(
+          [api](const EncodableValue& message,
+                const flutter::MessageReply<EncodableValue>& reply) {
+            try {
+              const auto& args = std::get<EncodableList>(message);
+              const auto& encodable_a_nullable_bool_arg = args.at(0);
+              const auto* a_nullable_bool_arg =
+                  std::get_if<bool>(&encodable_a_nullable_bool_arg);
+              const auto& encodable_a_nullable_int_arg = args.at(1);
+              const int64_t a_nullable_int_arg_value =
+                  encodable_a_nullable_int_arg.IsNull()
+                      ? 0
+                      : encodable_a_nullable_int_arg.LongValue();
+              const auto* a_nullable_int_arg =
+                  encodable_a_nullable_int_arg.IsNull()
+                      ? nullptr
+                      : &a_nullable_int_arg_value;
+              const auto& encodable_a_nullable_string_arg = args.at(2);
+              const auto* a_nullable_string_arg =
+                  std::get_if<std::string>(&encodable_a_nullable_string_arg);
+              api->CallFlutterSendMultipleNullableTypesWithoutRecursion(
+                  a_nullable_bool_arg, a_nullable_int_arg,
+                  a_nullable_string_arg,
+                  [reply](ErrorOr<AllNullableTypesWithoutRecursion>&& output) {
                     if (output.has_error()) {
                       reply(WrapError(output.error()));
                       return;
@@ -3705,9 +4562,13 @@ EncodableValue FlutterIntegrationCoreApiCodecSerializer::ReadValueOfType(
       return CustomEncodableValue(AllNullableTypes::FromEncodableList(
           std::get<EncodableList>(ReadValue(stream))));
     case 130:
+      return CustomEncodableValue(
+          AllNullableTypesWithoutRecursion::FromEncodableList(
+              std::get<EncodableList>(ReadValue(stream))));
+    case 131:
       return CustomEncodableValue(AllTypes::FromEncodableList(
           std::get<EncodableList>(ReadValue(stream))));
-    case 131:
+    case 132:
       return CustomEncodableValue(TestMessage::FromEncodableList(
           std::get<EncodableList>(ReadValue(stream))));
     default:
@@ -3734,15 +4595,23 @@ void FlutterIntegrationCoreApiCodecSerializer::WriteValue(
           stream);
       return;
     }
-    if (custom_value->type() == typeid(AllTypes)) {
+    if (custom_value->type() == typeid(AllNullableTypesWithoutRecursion)) {
       stream->WriteByte(130);
+      WriteValue(EncodableValue(std::any_cast<AllNullableTypesWithoutRecursion>(
+                                    *custom_value)
+                                    .ToEncodableList()),
+                 stream);
+      return;
+    }
+    if (custom_value->type() == typeid(AllTypes)) {
+      stream->WriteByte(131);
       WriteValue(EncodableValue(
                      std::any_cast<AllTypes>(*custom_value).ToEncodableList()),
                  stream);
       return;
     }
     if (custom_value->type() == typeid(TestMessage)) {
-      stream->WriteByte(131);
+      stream->WriteByte(132);
       WriteValue(
           EncodableValue(
               std::any_cast<TestMessage>(*custom_value).ToEncodableList()),
@@ -3970,6 +4839,88 @@ void FlutterIntegrationCoreApi::SendMultipleNullableTypes(
           } else {
             const auto& return_value = std::any_cast<const AllNullableTypes&>(
                 std::get<CustomEncodableValue>(list_return_value->at(0)));
+            on_success(return_value);
+          }
+        } else {
+          on_error(CreateConnectionError(channel_name));
+        }
+      });
+}
+
+void FlutterIntegrationCoreApi::EchoAllNullableTypesWithoutRecursion(
+    const AllNullableTypesWithoutRecursion* everything_arg,
+    std::function<void(const AllNullableTypesWithoutRecursion*)>&& on_success,
+    std::function<void(const FlutterError&)>&& on_error) {
+  const std::string channel_name =
+      "dev.flutter.pigeon.pigeon_integration_tests.FlutterIntegrationCoreApi."
+      "echoAllNullableTypesWithoutRecursion";
+  BasicMessageChannel<> channel(binary_messenger_, channel_name, &GetCodec());
+  EncodableValue encoded_api_arguments = EncodableValue(EncodableList{
+      everything_arg ? CustomEncodableValue(*everything_arg) : EncodableValue(),
+  });
+  channel.Send(
+      encoded_api_arguments, [channel_name, on_success = std::move(on_success),
+                              on_error = std::move(on_error)](
+                                 const uint8_t* reply, size_t reply_size) {
+        std::unique_ptr<EncodableValue> response =
+            GetCodec().DecodeMessage(reply, reply_size);
+        const auto& encodable_return_value = *response;
+        const auto* list_return_value =
+            std::get_if<EncodableList>(&encodable_return_value);
+        if (list_return_value) {
+          if (list_return_value->size() > 1) {
+            on_error(
+                FlutterError(std::get<std::string>(list_return_value->at(0)),
+                             std::get<std::string>(list_return_value->at(1)),
+                             list_return_value->at(2)));
+          } else {
+            const auto* return_value =
+                &(std::any_cast<const AllNullableTypesWithoutRecursion&>(
+                    std::get<CustomEncodableValue>(list_return_value->at(0))));
+            on_success(return_value);
+          }
+        } else {
+          on_error(CreateConnectionError(channel_name));
+        }
+      });
+}
+
+void FlutterIntegrationCoreApi::SendMultipleNullableTypesWithoutRecursion(
+    const bool* a_nullable_bool_arg, const int64_t* a_nullable_int_arg,
+    const std::string* a_nullable_string_arg,
+    std::function<void(const AllNullableTypesWithoutRecursion&)>&& on_success,
+    std::function<void(const FlutterError&)>&& on_error) {
+  const std::string channel_name =
+      "dev.flutter.pigeon.pigeon_integration_tests.FlutterIntegrationCoreApi."
+      "sendMultipleNullableTypesWithoutRecursion";
+  BasicMessageChannel<> channel(binary_messenger_, channel_name, &GetCodec());
+  EncodableValue encoded_api_arguments = EncodableValue(EncodableList{
+      a_nullable_bool_arg ? EncodableValue(*a_nullable_bool_arg)
+                          : EncodableValue(),
+      a_nullable_int_arg ? EncodableValue(*a_nullable_int_arg)
+                         : EncodableValue(),
+      a_nullable_string_arg ? EncodableValue(*a_nullable_string_arg)
+                            : EncodableValue(),
+  });
+  channel.Send(
+      encoded_api_arguments, [channel_name, on_success = std::move(on_success),
+                              on_error = std::move(on_error)](
+                                 const uint8_t* reply, size_t reply_size) {
+        std::unique_ptr<EncodableValue> response =
+            GetCodec().DecodeMessage(reply, reply_size);
+        const auto& encodable_return_value = *response;
+        const auto* list_return_value =
+            std::get_if<EncodableList>(&encodable_return_value);
+        if (list_return_value) {
+          if (list_return_value->size() > 1) {
+            on_error(
+                FlutterError(std::get<std::string>(list_return_value->at(0)),
+                             std::get<std::string>(list_return_value->at(1)),
+                             list_return_value->at(2)));
+          } else {
+            const auto& return_value =
+                std::any_cast<const AllNullableTypesWithoutRecursion&>(
+                    std::get<CustomEncodableValue>(list_return_value->at(0)));
             on_success(return_value);
           }
         } else {
