@@ -41,6 +41,18 @@ import 'table_span.dart';
 /// the [TableCellDelegateMixin]. The [TableView.builder] and [TableView.list]
 /// constructors create their own delegate.
 ///
+/// A table with infinite rows and columns can be made by using a
+/// [TableCellBuilderDelegate], or the [TableView.builder] constructor, and
+/// omitting the row or column count. Returning null from the
+/// [columnBuilder] or [rowBuilder] in this case will terminate
+/// the row or column at that index, representing the end of the table in that
+/// axis. In this scenario, until the potential end of the table in either
+/// dimension is reached by returning null, the
+/// [ScrollPosition.maxScrollExtent] will reflect [double.infinity]. This is
+/// because as the table is built lazily, it will not know the end has been
+/// reached until the [ScrollPosition] arrives there. This is similar to
+/// returning null from [ListView.builder] to signify the end of the list.
+///
 /// This example shows a TableView of 100 children, all sized 100 by 100
 /// pixels with a few [TableSpanDecoration]s like background colors and borders.
 /// The `builder` constructor is called on demand for the cells that are visible
@@ -116,6 +128,16 @@ class TableView extends TwoDimensionalScrollView {
   /// This constructor generates a [TableCellBuilderDelegate] for building
   /// children on demand using the required [cellBuilder],
   /// [columnBuilder], and [rowBuilder].
+  ///
+  /// For infinite rows and columns, omit providing [columnCount] or [rowCount].
+  /// Returning null from the [columnBuilder] or [rowBuilder] will terminate
+  /// the row or column at that index, representing the end of the table in that
+  /// axis. In this scenario, until the potential end of the table in either
+  /// dimension is reached by returning null, the
+  /// [ScrollPosition.maxScrollExtent] will reflect [double.infinity]. This is
+  /// because as the table is built lazily, it will not know the end has been
+  /// reached until the [ScrollPosition] arrives there. This is similar to
+  /// returning null from [ListView.builder] to signify the end of the list.
   TableView.builder({
     super.key,
     super.primary,
@@ -129,17 +151,17 @@ class TableView extends TwoDimensionalScrollView {
     super.clipBehavior,
     int pinnedRowCount = 0,
     int pinnedColumnCount = 0,
-    required int columnCount,
-    required int rowCount,
+    int? columnCount,
+    int? rowCount,
     required TableSpanBuilder columnBuilder,
     required TableSpanBuilder rowBuilder,
     required TableViewCellBuilder cellBuilder,
   })  : assert(pinnedRowCount >= 0),
-        assert(rowCount >= 0),
-        assert(rowCount >= pinnedRowCount),
-        assert(columnCount >= 0),
+        assert(rowCount == null || rowCount >= 0),
+        assert(rowCount == null || rowCount >= pinnedRowCount),
+        assert(columnCount == null || columnCount >= 0),
         assert(pinnedColumnCount >= 0),
-        assert(columnCount >= pinnedColumnCount),
+        assert(columnCount == null || columnCount >= pinnedColumnCount),
         super(
           delegate: TableCellBuilderDelegate(
             columnCount: columnCount,
@@ -302,12 +324,32 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
   final List<int> _mergedColumns = <int>[];
 
   // Cached Table metrics
-  Map<int, _Span> _columnMetrics = <int, _Span>{};
-  Map<int, _Span> _rowMetrics = <int, _Span>{};
+  final Map<int, _Span> _columnMetrics = <int, _Span>{};
+  final Map<int, _Span> _rowMetrics = <int, _Span>{};
   int? _firstNonPinnedRow;
   int? _firstNonPinnedColumn;
   int? _lastNonPinnedRow;
   int? _lastNonPinnedColumn;
+
+  int? _columnNullTerminatedIndex;
+  bool get _columnsAreInfinite => delegate.columnCount == null;
+  // How far columns should be laid out in a given frame.
+  double get _targetColumnPixel {
+    return cacheExtent +
+        horizontalOffset.pixels +
+        viewportDimension.width -
+        _pinnedColumnsExtent;
+  }
+
+  int? _rowNullTerminatedIndex;
+  bool get _rowsAreInfinite => delegate.rowCount == null;
+  // How far rows should be laid out in a given frame.
+  double get _targetRowPixel {
+    return cacheExtent +
+        verticalOffset.pixels +
+        viewportDimension.height -
+        _pinnedRowsExtent;
+  }
 
   TableVicinity? get _firstNonPinnedCell {
     if (_firstNonPinnedRow == null || _firstNonPinnedColumn == null) {
@@ -406,31 +448,79 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
     return false;
   }
 
-  // Updates the cached metrics for the table.
+  // Updates the cached column metrics for the table.
   //
-  // Will iterate through all columns and rows to define the layout pattern of
-  // the cells of the table.
-  //
-  // TODO(Piinks): Add back infinite separately for easier review, https://github.com/flutter/flutter/issues/131226
-  // Only relevant when the number of rows and columns is finite
-  void _updateAllMetrics() {
-    assert(needsDelegateRebuild || didResize);
+  // By default, existing column metrics will be updated if they have changed.
+  // Setting `appendColumns` to false will preserve existing metrics, adding
+  // additional metrics up to the visible+cacheExtent or up to a provided index,
+  // `toColumnIndex`. Appending is only relevant when the number of columns is
+  // infinite.
+  void _updateColumnMetrics({bool appendColumns = false, int? toColumnIndex}) {
+    assert(() {
+      if (toColumnIndex != null) {
+        // If we are computing up to an index, we must be appending.
+        return appendColumns;
+      }
+      return true;
+    }());
+    double startOfRegularColumn = 0.0;
+    double startOfPinnedColumn = 0.0;
+    if (appendColumns) {
+      // We are only adding to the metrics we already know, since we are lazily
+      // compiling metrics. This should only be the case when the
+      // number of columns is infinite, and saves us going through all the
+      // columns we already know about.
+      assert(_columnsAreInfinite);
+      assert(_columnMetrics.isNotEmpty);
+      startOfPinnedColumn =
+          _columnMetrics[_firstNonPinnedColumn]?.trailingOffset ?? 0.0;
+      startOfRegularColumn =
+          _columnMetrics[_lastNonPinnedColumn]?.trailingOffset ?? 0.0;
+    }
+    // If we are computing up to a specific index, we are getting info for a
+    // merged cell, do not change the visible cells.
+    _firstNonPinnedColumn =
+        toColumnIndex == null ? null : _firstNonPinnedColumn;
+    _lastNonPinnedColumn = toColumnIndex == null ? null : _lastNonPinnedColumn;
+    int column = appendColumns ? _columnMetrics.length : 0;
 
-    _firstNonPinnedColumn = null;
-    _lastNonPinnedColumn = null;
-    double startOfRegularColumn = 0;
-    double startOfPinnedColumn = 0;
+    bool reachedColumnEnd() {
+      if (_columnsAreInfinite) {
+        if (toColumnIndex != null) {
+          // Column metrics should be computed up to the provided index.
+          // Only relevant when we are filling in missing column metrics in an
+          // infinite context.
+          return _columnMetrics.length > toColumnIndex;
+        }
+        // There are infinite columns, and no target index, compute metrics
+        // up to what is visible and in the cache extent, or the index that null
+        // terminates.
+        return _lastNonPinnedColumn != null ||
+            _columnNullTerminatedIndex != null;
+      }
+      // Compute all the metrics if the columns are finite.
+      return column == delegate.columnCount!;
+    }
 
-    final Map<int, _Span> newColumnMetrics = <int, _Span>{};
-    for (int column = 0; column < delegate.columnCount; column++) {
+    while (!reachedColumnEnd()) {
       final bool isPinned = column < delegate.pinnedColumnCount;
       final double leadingOffset =
           isPinned ? startOfPinnedColumn : startOfRegularColumn;
       _Span? span = _columnMetrics.remove(column);
-      assert(needsDelegateRebuild || span != null);
-      final TableSpan configuration = needsDelegateRebuild
-          ? delegate.buildColumn(column)
-          : span!.configuration;
+      final TableSpan? configuration =
+          span?.configuration ?? delegate.buildColumn(column);
+      if (configuration == null) {
+        // We have reached the end of columns based on a null termination. This
+        // This happens when a column count has not been specified.
+        assert(_columnsAreInfinite);
+        _lastNonPinnedColumn ??= column - 1;
+        _columnNullTerminatedIndex = column;
+        final bool acceptedDimension = _updateHorizontalScrollBounds();
+        if (!acceptedDimension) {
+          _updateFirstAndLastVisibleCell();
+        }
+        break;
+      }
       span ??= _Span();
       span.update(
         isPinned: isPinned,
@@ -443,17 +533,13 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
           ),
         ),
       );
-      newColumnMetrics[column] = span;
+      _columnMetrics[column] = span;
       if (!isPinned) {
         if (span.trailingOffset >= horizontalOffset.pixels &&
             _firstNonPinnedColumn == null) {
           _firstNonPinnedColumn = column;
         }
-        final double targetColumnPixel = cacheExtent +
-            horizontalOffset.pixels +
-            viewportDimension.width -
-            startOfPinnedColumn;
-        if (span.trailingOffset >= targetColumnPixel &&
+        if (span.trailingOffset >= _targetColumnPixel &&
             _lastNonPinnedColumn == null) {
           _lastNonPinnedColumn = column;
         }
@@ -461,27 +547,82 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
       } else {
         startOfPinnedColumn = span.trailingOffset;
       }
+      column++;
     }
-    assert(newColumnMetrics.length >= delegate.pinnedColumnCount);
-    for (final _Span span in _columnMetrics.values) {
-      span.dispose();
+
+    assert(_columnMetrics.length >= delegate.pinnedColumnCount);
+  }
+
+  // Updates the cached row metrics for the table.
+  //
+  // By default, existing row metrics will be updated if they have changed.
+  // Setting `appendRows` to false will preserve existing metrics, adding
+  // additional metrics up to the visible+cacheExtent or up to a provided index,
+  // `toRowIndex`. Appending is only relevant when the number of rows is
+  // infinite.
+  void _updateRowMetrics({bool appendRows = false, int? toRowIndex}) {
+    assert(() {
+      if (toRowIndex != null) {
+        // If we are computing up to an index, we must be appending.
+        return appendRows;
+      }
+      return true;
+    }());
+    double startOfRegularRow = 0.0;
+    double startOfPinnedRow = 0.0;
+    if (appendRows) {
+      // We are only adding to the metrics we already know, since we are lazily
+      // compiling metrics. This should only be the case when the
+      // number of rows is infinite, and saves us going through all the
+      // rows we already know about.
+      assert(_rowsAreInfinite);
+      assert(_rowMetrics.isNotEmpty);
+      startOfPinnedRow = _rowMetrics[_firstNonPinnedRow]?.trailingOffset ?? 0.0;
+      startOfRegularRow = _rowMetrics[_lastNonPinnedRow]?.trailingOffset ?? 0.0;
     }
-    _columnMetrics = newColumnMetrics;
+    // If we are computing up to a specific index, we are getting info for a
+    // merged cell, do not change the visible cells.
+    _firstNonPinnedRow = toRowIndex == null ? null : _firstNonPinnedRow;
+    _lastNonPinnedRow = toRowIndex == null ? null : _lastNonPinnedRow;
+    int row = appendRows ? _rowMetrics.length : 0;
 
-    _firstNonPinnedRow = null;
-    _lastNonPinnedRow = null;
-    double startOfRegularRow = 0;
-    double startOfPinnedRow = 0;
+    bool reachedRowEnd() {
+      if (_rowsAreInfinite) {
+        if (toRowIndex != null) {
+          // Row metrics should be computed up to the provided index.
+          // Only relevant when we are filling in missing column metrics in an
+          // infinite context.
+          return _rowMetrics.length > toRowIndex;
+        }
+        // There are infinite row, and no target index, compute metrics
+        // up to what is visible and in the cache extent, or the index that null
+        // terminates.
+        return _lastNonPinnedRow != null || _rowNullTerminatedIndex != null;
+      }
+      // Compute all the metrics if the rows are finite.
+      return row == delegate.rowCount!;
+    }
 
-    final Map<int, _Span> newRowMetrics = <int, _Span>{};
-    for (int row = 0; row < delegate.rowCount; row++) {
+    while (!reachedRowEnd()) {
       final bool isPinned = row < delegate.pinnedRowCount;
       final double leadingOffset =
           isPinned ? startOfPinnedRow : startOfRegularRow;
       _Span? span = _rowMetrics.remove(row);
-      assert(needsDelegateRebuild || span != null);
-      final TableSpan configuration =
-          needsDelegateRebuild ? delegate.buildRow(row) : span!.configuration;
+      final TableSpan? configuration =
+          span?.configuration ?? delegate.buildRow(row);
+      if (configuration == null) {
+        // We have reached the end of rows based on a null termination. This
+        // This happens when a row count has not been specified, but we have
+        // reached the end.
+        assert(_rowsAreInfinite);
+        _lastNonPinnedRow ??= row - 1;
+        _rowNullTerminatedIndex = row;
+        final bool acceptedDimension = _updateVerticalScrollBounds();
+        if (!acceptedDimension) {
+          _updateFirstAndLastVisibleCell();
+        }
+        break;
+      }
       span ??= _Span();
       span.update(
         isPinned: isPinned,
@@ -494,17 +635,13 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
           ),
         ),
       );
-      newRowMetrics[row] = span;
+      _rowMetrics[row] = span;
       if (!isPinned) {
         if (span.trailingOffset >= verticalOffset.pixels &&
             _firstNonPinnedRow == null) {
           _firstNonPinnedRow = row;
         }
-        final double targetRowPixel = cacheExtent +
-            verticalOffset.pixels +
-            viewportDimension.height -
-            startOfPinnedRow;
-        if (span.trailingOffset >= targetRowPixel &&
+        if (span.trailingOffset > _targetRowPixel &&
             _lastNonPinnedRow == null) {
           _lastNonPinnedRow = row;
         }
@@ -512,32 +649,26 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
       } else {
         startOfPinnedRow = span.trailingOffset;
       }
-    }
-    assert(newRowMetrics.length >= delegate.pinnedRowCount);
-    for (final _Span span in _rowMetrics.values) {
-      span.dispose();
-    }
-    _rowMetrics = newRowMetrics;
-
-    final double maxVerticalScrollExtent;
-    if (_rowMetrics.length <= delegate.pinnedRowCount) {
-      assert(_firstNonPinnedRow == null && _lastNonPinnedRow == null);
-      maxVerticalScrollExtent = 0.0;
-    } else {
-      final int lastRow = _rowMetrics.length - 1;
-      if (_firstNonPinnedRow != null) {
-        _lastNonPinnedRow ??= lastRow;
-      }
-      maxVerticalScrollExtent = math.max(
-        0.0,
-        _rowMetrics[lastRow]!.trailingOffset -
-            viewportDimension.height +
-            startOfPinnedRow,
-      );
+      row++;
     }
 
+    assert(_rowMetrics.length >= delegate.pinnedRowCount);
+  }
+
+  void _updateScrollBounds() {
+    final bool acceptedDimension =
+        _updateHorizontalScrollBounds() && _updateVerticalScrollBounds();
+    if (!acceptedDimension) {
+      _updateFirstAndLastVisibleCell();
+    }
+  }
+
+  bool _updateHorizontalScrollBounds() {
     final double maxHorizontalScrollExtent;
-    if (_columnMetrics.length <= delegate.pinnedColumnCount) {
+    if (_columnsAreInfinite && _columnNullTerminatedIndex == null) {
+      maxHorizontalScrollExtent = double.infinity;
+    } else if (!_columnsAreInfinite &&
+        _columnMetrics.length <= delegate.pinnedColumnCount) {
       assert(_firstNonPinnedColumn == null && _lastNonPinnedColumn == null);
       maxHorizontalScrollExtent = 0.0;
     } else {
@@ -549,29 +680,61 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
         0.0,
         _columnMetrics[lastColumn]!.trailingOffset -
             viewportDimension.width +
-            startOfPinnedColumn,
+            _pinnedColumnsExtent,
       );
     }
-
-    final bool acceptedDimension = horizontalOffset.applyContentDimensions(
-            0.0, maxHorizontalScrollExtent) &&
-        verticalOffset.applyContentDimensions(0.0, maxVerticalScrollExtent);
-    if (!acceptedDimension) {
-      _updateFirstAndLastVisibleCell();
-    }
+    return horizontalOffset.applyContentDimensions(
+      0.0,
+      maxHorizontalScrollExtent,
+    );
   }
 
-  // Uses the cached metrics to update the currently visible cells
-  //
-  // TODO(Piinks): Add back infinite separately for easier review, https://github.com/flutter/flutter/issues/131226
-  // Only relevant when the number of rows and columns is finite
+  bool _updateVerticalScrollBounds() {
+    final double maxVerticalScrollExtent;
+    if (_rowsAreInfinite && _rowNullTerminatedIndex == null) {
+      maxVerticalScrollExtent = double.infinity;
+    } else if (!_rowsAreInfinite &&
+        _rowMetrics.length <= delegate.pinnedRowCount) {
+      assert(_firstNonPinnedRow == null && _lastNonPinnedRow == null);
+      maxVerticalScrollExtent = 0.0;
+    } else {
+      final int lastRow = _rowMetrics.length - 1;
+      if (_firstNonPinnedRow != null) {
+        _lastNonPinnedRow ??= lastRow;
+      }
+      maxVerticalScrollExtent = math.max(
+        0.0,
+        _rowMetrics[lastRow]!.trailingOffset -
+            viewportDimension.height +
+            _pinnedRowsExtent,
+      );
+    }
+    return verticalOffset.applyContentDimensions(
+      0.0,
+      maxVerticalScrollExtent,
+    );
+  }
+
+  // Uses the cached metrics to update the currently visible cells. If the
+  // number of rows or columns are infinite, the layout is computed lazily, so
+  // this will call for an update to the metrics if we have scrolled beyond the
+  // layout portion we know about.
   void _updateFirstAndLastVisibleCell() {
+    if (_columnMetrics.isNotEmpty) {
+      _Span lastKnownColumn = _columnMetrics[_columnMetrics.length - 1]!;
+      if (_columnsAreInfinite &&
+          lastKnownColumn.trailingOffset < _targetColumnPixel) {
+        // This will add the column metrics we do not know about up to the
+        // _targetColumnPixel, while keeping the ones we already know about.
+        _updateColumnMetrics(appendColumns: true);
+        lastKnownColumn = _columnMetrics[_columnMetrics.length - 1]!;
+        assert(_columnMetrics.length == delegate.columnCount ||
+            lastKnownColumn.trailingOffset >= _targetColumnPixel ||
+            _columnNullTerminatedIndex != null);
+      }
+    }
     _firstNonPinnedColumn = null;
     _lastNonPinnedColumn = null;
-    final double targetColumnPixel = cacheExtent +
-        horizontalOffset.pixels +
-        viewportDimension.width -
-        _pinnedColumnsExtent;
     for (int column = 0; column < _columnMetrics.length; column++) {
       if (_columnMetrics[column]!.isPinned) {
         continue;
@@ -581,7 +744,7 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
           _firstNonPinnedColumn == null) {
         _firstNonPinnedColumn = column;
       }
-      if (endOfColumn >= targetColumnPixel && _lastNonPinnedColumn == null) {
+      if (endOfColumn >= _targetColumnPixel && _lastNonPinnedColumn == null) {
         _lastNonPinnedColumn = column;
         break;
       }
@@ -590,12 +753,20 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
       _lastNonPinnedColumn ??= _columnMetrics.length - 1;
     }
 
+    if (_rowMetrics.isNotEmpty) {
+      _Span lastKnownRow = _rowMetrics[_rowMetrics.length - 1]!;
+      if (_rowsAreInfinite && lastKnownRow.trailingOffset < _targetRowPixel) {
+        // This will add the row metrics we do not know about up to the
+        // _targetRowPixel, while keeping the ones we already know about.
+        _updateRowMetrics(appendRows: true);
+        lastKnownRow = _rowMetrics[_rowMetrics.length - 1]!;
+        assert(_rowMetrics.length == delegate.rowCount ||
+            lastKnownRow.trailingOffset >= _targetRowPixel ||
+            _rowNullTerminatedIndex != null);
+      }
+    }
     _firstNonPinnedRow = null;
     _lastNonPinnedRow = null;
-    final double targetRowPixel = cacheExtent +
-        verticalOffset.pixels +
-        viewportDimension.height -
-        _pinnedRowsExtent;
     for (int row = 0; row < _rowMetrics.length; row++) {
       if (_rowMetrics[row]!.isPinned) {
         continue;
@@ -604,7 +775,7 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
       if (endOfRow >= verticalOffset.pixels && _firstNonPinnedRow == null) {
         _firstNonPinnedRow = row;
       }
-      if (endOfRow >= targetRowPixel && _lastNonPinnedRow == null) {
+      if (endOfRow >= _targetRowPixel && _lastNonPinnedRow == null) {
         _lastNonPinnedRow = row;
         break;
       }
@@ -617,13 +788,27 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
   @override
   void layoutChildSequence() {
     // Reset for a new frame
+    // We always reset the null terminating indices in case rows or columns have
+    // been added or removed.
+    _rowNullTerminatedIndex = null;
+    _columnNullTerminatedIndex = null;
     _mergedVicinities.clear();
     _mergedRows.clear();
     _mergedColumns.clear();
 
     if (needsDelegateRebuild || didResize) {
       // Recomputes the table metrics, invalidates any cached information.
-      _updateAllMetrics();
+      for (final _Span span in _columnMetrics.values) {
+        span.dispose();
+      }
+      _columnMetrics.clear();
+      for (final _Span span in _rowMetrics.values) {
+        span.dispose();
+      }
+      _rowMetrics.clear();
+      _updateColumnMetrics();
+      _updateRowMetrics();
+      _updateScrollBounds();
     } else {
       // Updates the visible cells based on cached table metrics.
       _updateFirstAndLastVisibleCell();
@@ -696,7 +881,7 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
     required int currentSpan,
     required int spanMergeStart,
     required int spanMergeEnd,
-    required int spanCount,
+    required int? spanCount,
     required int pinnedSpanCount,
     required TableVicinity currentVicinity,
   }) {
@@ -712,7 +897,7 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
       'than the current $lowerSpanOrientation at $currentVicinity.',
     );
     assert(
-      spanMergeEnd < spanCount,
+      spanCount == null || spanMergeEnd < spanCount,
       '$spanOrientation merge configuration exceeds number of '
       '${lowerSpanOrientation}s in the table. $spanOrientation merge '
       'containing $currentVicinity starts at $spanMergeStart, and ends at '
@@ -740,6 +925,7 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
     double rowOffset = -offset.dy;
     for (int row = start.row; row <= end.row; row += 1) {
       double columnOffset = -offset.dx;
+      assert(row < _rowMetrics.length);
       rowSpan = _rowMetrics[row]!;
       final double standardRowHeight = rowSpan.extent;
       double? mergedRowHeight;
@@ -747,6 +933,7 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
       rowOffset += rowSpan.configuration.padding.leading;
 
       for (int column = start.column; column <= end.column; column += 1) {
+        assert(column < _columnMetrics.length);
         colSpan = _columnMetrics[column]!;
         final double standardColumnWidth = colSpan.extent;
         double? mergedColumnWidth;
@@ -819,6 +1006,19 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
             mergedRowOffset = baseRowOffset +
                 _rowMetrics[firstRow]!.leadingOffset +
                 _rowMetrics[firstRow]!.configuration.padding.leading;
+            if (_rowsAreInfinite && _rowMetrics[lastRow] == null) {
+              // The number of rows is infinte, and we have not calculated
+              // the metrics to the full extent of the merged cell. Update the
+              // metrics so we have all the information for the merged area.
+              _updateRowMetrics(appendRows: true, toRowIndex: lastRow);
+            }
+            assert(
+              _rowMetrics[lastRow] != null,
+              'The merged cell containing $vicinity is missing TableSpan '
+              'information necessary for layout. The rowBuilder returned '
+              'null, signifying the end, at row $_rowNullTerminatedIndex but the '
+              'merged cell is configured to end with row $lastRow.',
+            );
             mergedRowHeight = _rowMetrics[lastRow]!.trailingOffset -
                 _rowMetrics[firstRow]!.leadingOffset -
                 _rowMetrics[lastRow]!.configuration.padding.trailing -
@@ -840,6 +1040,23 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
             mergedColumnOffset = baseColumnOffset +
                 _columnMetrics[firstColumn]!.leadingOffset +
                 _columnMetrics[firstColumn]!.configuration.padding.leading;
+
+            if (_columnsAreInfinite && _columnMetrics[lastColumn] == null) {
+              // The number of columns is infinte, and we have not calculated
+              // the metrics to the full extent of the merged cell. Update the
+              // metrics so we have all the information for the merged area.
+              _updateColumnMetrics(
+                appendColumns: true,
+                toColumnIndex: lastColumn,
+              );
+            }
+            assert(
+              _columnMetrics[lastColumn] != null,
+              'The merged cell containing $vicinity is missing TableSpan '
+              'information necessary for layout. The columnBuilder returned '
+              'null, signifying the end, at column $_columnNullTerminatedIndex but '
+              'the merged cell is configured to end with column $lastColumn.',
+            );
             mergedColumnWidth = _columnMetrics[lastColumn]!.trailingOffset -
                 _columnMetrics[firstColumn]!.leadingOffset -
                 _columnMetrics[lastColumn]!.configuration.padding.trailing -
@@ -1040,7 +1257,10 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
     // A merged cell spans multiple vicinities, but only lays out one child for
     // the full area. Returns the child that has been laid out to span the given
     // vicinity.
-    assert(_mergedVicinities.keys.contains(vicinity));
+    assert(
+      _mergedVicinities.keys.contains(vicinity),
+      'The vicinity $vicinity is not accounted for as covered by a merged cell.',
+    );
     final TableVicinity mergedVicinity = _mergedVicinities[vicinity]!;
     // This vicinity must resolve to a child, unless something has gone wrong!
     return getChildFor(
@@ -1468,6 +1688,12 @@ class RenderTableViewport extends RenderTwoDimensionalViewport {
     _clipPinnedRowsHandle.layer = null;
     _clipPinnedColumnsHandle.layer = null;
     _clipCellsHandle.layer = null;
+    for (final _Span span in _rowMetrics.values) {
+      span.dispose();
+    }
+    for (final _Span span in _columnMetrics.values) {
+      span.dispose();
+    }
     super.dispose();
   }
 }
