@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
@@ -180,6 +181,1861 @@ void main() {
         ),
       );
       expect(tableView, isNull);
+    });
+
+    group('Infinite spans - ', () {
+      late ScrollController verticalController;
+      late ScrollController horizontalController;
+      const TableSpan largeSpan = TableSpan(extent: FixedTableSpanExtent(200));
+
+      setUp(() {
+        verticalController = ScrollController();
+        horizontalController = ScrollController();
+      });
+
+      tearDown(() {
+        verticalController.dispose();
+        horizontalController.dispose();
+      });
+
+      TableView getTableView({
+        int? columnCount,
+        int? rowCount,
+        TableSpanBuilder? columnBuilder,
+        TableSpanBuilder? rowBuilder,
+        TableViewCellBuilder? cellBuilder,
+        int pinnedColumnCount = 0,
+        int pinnedRowCount = 0,
+      }) {
+        return TableView.builder(
+          verticalDetails: ScrollableDetails.vertical(
+            controller: verticalController,
+          ),
+          horizontalDetails: ScrollableDetails.horizontal(
+            controller: horizontalController,
+          ),
+          columnCount: columnCount,
+          pinnedColumnCount: pinnedColumnCount,
+          columnBuilder: columnBuilder ?? (_) => largeSpan,
+          rowCount: rowCount,
+          pinnedRowCount: pinnedRowCount,
+          rowBuilder: rowBuilder ?? (_) => largeSpan,
+          cellBuilder: cellBuilder ??
+              (_, TableVicinity vicinity) {
+                return TableViewCell(
+                  child: Text('R${vicinity.row}:C${vicinity.column}'),
+                );
+              },
+        );
+      }
+
+      testWidgets('infinite rows, columns are finite',
+          (WidgetTester tester) async {
+        // Nothing pinned ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(columnCount: 10),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out before row 5, or after row 9.
+        expect(find.text('R0:C0'), findsNothing);
+        expect(find.text('R10:C0'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            pinnedColumnCount: 1,
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out before row 5, or after row 9.
+        expect(find.text('R0:C0'), findsNothing);
+        expect(find.text('R10:C0'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            pinnedRowCount: 1,
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out before row 5, or after row 9, except for pinned row
+        // 0.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R1:C0'), findsNothing);
+        expect(find.text('R10:C0'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns and rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            pinnedColumnCount: 1,
+            pinnedRowCount: 1,
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out before row 5, or after row 9, except for pinned row
+        // 0.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R1:C0'), findsNothing);
+        expect(find.text('R10:C0'), findsNothing);
+      });
+
+      testWidgets('infinite columns, rows are finite',
+          (WidgetTester tester) async {
+        // Nothing pinned ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(rowCount: 10),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12.
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C12'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(rowCount: 10, pinnedColumnCount: 1),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12, except for
+        // pinned column.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C12'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(rowCount: 10, pinnedRowCount: 1),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12.
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C12'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns and rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            pinnedRowCount: 1,
+            pinnedColumnCount: 1,
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12, except for
+        // pinned column.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C12'), findsNothing);
+      });
+
+      testWidgets('infinite rows & columns', (WidgetTester tester) async {
+        // No pinned ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5, no rows beyond row 4.
+        expect(find.text('R0:C6'), findsNothing);
+        expect(find.text('R5:C0'), findsNothing);
+        // Change both scroll offsets, validate more columns and rows were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12.
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C12'), findsNothing);
+        // No rows laid out before row 4, or after row 9.
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R10:C6'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(pinnedColumnCount: 1),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5, no rows beyond row 4.
+        expect(find.text('R0:C6'), findsNothing);
+        expect(find.text('R5:C0'), findsNothing);
+        // Change both scroll offsets, validate more columns and rows were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12, except for
+        // pinned first column.
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C12'), findsNothing);
+        // No rows laid out before row 4, or after row 9.
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R10:C6'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned Rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(pinnedRowCount: 1),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5, no rows beyond row 4.
+        expect(find.text('R0:C6'), findsNothing);
+        expect(find.text('R5:C0'), findsNothing);
+        // Change both scroll offsets, validate more columns and rows were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12.
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C12'), findsNothing);
+        // No rows laid out before row 4, or after row 9, except for pinned
+        // first row.
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R10:C6'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns and rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            pinnedRowCount: 1,
+            pinnedColumnCount: 1,
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5, no rows beyond row 4.
+        expect(find.text('R0:C6'), findsNothing);
+        expect(find.text('R5:C0'), findsNothing);
+        // Change both scroll offsets, validate more columns and rows were
+        // populated.
+        horizontalController.jumpTo(1200.0);
+        verticalController.jumpTo(1000.0);
+        await tester.pump();
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R9:C11'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R9:C11')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 12, except for
+        // pinned first column.
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C12'), findsNothing);
+        // No rows laid out before row 4, or after row 9, except for pinned
+        // first row.
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R10:C6'), findsNothing);
+      });
+
+      testWidgets('infinite rows can null terminate',
+          (WidgetTester tester) async {
+        // Nothing pinned ---
+        bool calledOutOfBounds = false;
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              if (index > 8) {
+                calledOutOfBounds = true;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        // This exceeds the bounds of the scroll view once the rows have been
+        // null terminated.
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // After null terminating, the builder was not called further.
+        expect(calledOutOfBounds, isFalse);
+        // Max scroll extent was updated to reflect reaching the end of the rows
+        // after returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C4')),
+          const Rect.fromLTRB(800.0, 400.0, 1000.0, 600.0),
+        );
+        // No rows laid out before row 5, or after row 7.
+        expect(find.text('R0:C0'), findsNothing);
+        expect(find.text('R8:C0'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            pinnedColumnCount: 1,
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        // This exceeds the bounds of the scroll view once the rows have been
+        // null terminated.
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // Max scroll extent was updated to reflect reaching the end of the rows
+        // after returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C4')),
+          const Rect.fromLTRB(800.0, 400.0, 1000.0, 600.0),
+        );
+        // No rows laid out before row 5, or after row 7.
+        expect(find.text('R0:C0'), findsNothing);
+        expect(find.text('R8:C0'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            pinnedRowCount: 1,
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        // This exceeds the bounds of the scroll view once the rows have been
+        // null terminated.
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // Max scroll extent was updated to reflect reaching the end of the rows
+        // after returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C4')),
+          const Rect.fromLTRB(800.0, 400.0, 1000.0, 600.0),
+        );
+        // No rows laid out before row 5, or after row 7, except for the first
+        // pinned row.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R1:C0'), findsNothing);
+        expect(find.text('R8:C0'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns and rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            pinnedColumnCount: 1,
+            pinnedRowCount: 1,
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 800.0, 1000.0, 1000.0),
+        );
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the vertical scroll offset, validate more rows were populated.
+        // This exceeds the bounds of the scroll view once the rows have been
+        // null terminated.
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // Max scroll extent was updated to reflect reaching the end of the rows
+        // after returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C4')),
+          const Rect.fromLTRB(800.0, 400.0, 1000.0, 600.0),
+        );
+        // No rows laid out before row 5, or after row 7, except for the first
+        // pinned row.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R1:C0'), findsNothing);
+        expect(find.text('R8:C0'), findsNothing);
+      });
+
+      testWidgets('Null terminated rows will update',
+          (WidgetTester tester) async {
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        // Change the vertical scroll offset, validate more rows were populated.
+        // This exceeds the bounds of the scroll view once the rows have been
+        // null terminated.
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // Max scroll extent was updated to reflect reaching the end of the rows
+        // after returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C4')),
+          const Rect.fromLTRB(800.0, 400.0, 1000.0, 600.0),
+        );
+        // No rows laid out before row 5, or after row 7.
+        expect(find.text('R0:C0'), findsNothing);
+        expect(find.text('R8:C0'), findsNothing);
+
+        // Increase the number of rows
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            rowBuilder: (int index) {
+              // There will only be 16 rows.
+              if (index == 16) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+
+        // The position should not have changed.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // Max scroll extent was updated to reflect we no longer know where the
+        // end is, until the rowBuilder returns null again.
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        // The layout should not have changed.
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C4')),
+          const Rect.fromLTRB(800.0, 400.0, 1000.0, 600.0),
+        );
+        // No rows laid out before row 5, but more rows were laid out into the
+        // newly updated cacheExtent (row 8).
+        expect(find.text('R0:C0'), findsNothing);
+        expect(find.text('R8:C0'), findsOneWidget);
+        // This exceeds the new bounds.
+        verticalController.jumpTo(3200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 2600.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // Max scroll extent was updated to reflect reaching the end of the rows
+        // after returning null again at the new index.
+        expect(verticalController.position.maxScrollExtent, 2600.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+
+        // Decrease the number of rows
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnCount: 10,
+            rowBuilder: (int index) {
+              // There will only be 5 rows.
+              if (index == 5) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+
+        // The position should have changed.
+        expect(verticalController.position.pixels, 400.0);
+        expect(horizontalController.position.pixels, 0.0);
+        // Max scroll extent was updated to the new end we have corrected to.
+        expect(verticalController.position.maxScrollExtent, 400.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        // The layout updated.
+        expect(find.text('R2:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R2:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(800.0, 400.0, 1000.0, 600.0),
+        );
+        // No rows laid out after row 5.
+        expect(find.text('R5:C0'), findsNothing);
+      });
+
+      testWidgets('Null terminated columns will update',
+          (WidgetTester tester) async {
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            columnBuilder: (int index) {
+              // There will only be 8 columns.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        // Change the horizontal scroll offset, validate more columns were
+        // populated. This exceeds the bounds of the scroll view once the
+        // columns have been null terminated.
+        horizontalController.jumpTo(1400.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 800.0);
+        // Max scroll extent was updated to reflect reaching the end of the
+        // columns after returning null.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, 800.0);
+        expect(find.text('R0:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C5')),
+          const Rect.fromLTRB(200.0, 0.0, 400.0, 200.0),
+        );
+        expect(find.text('R4:C7'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C7')),
+          const Rect.fromLTRB(600.0, 800.0, 800.0, 1000.0),
+        );
+        // No columns laid out before column 3, or after column 7.
+        expect(find.text('R0:C2'), findsNothing);
+        expect(find.text('R0:C8'), findsNothing);
+
+        // Increase the number of rows
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            columnBuilder: (int index) {
+              // There will only be 16 column.
+              if (index == 16) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+
+        // The position should not have changed.
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 800.0);
+        // Max scroll extent was updated to reflect we no longer know where the
+        // end is, until the rowBuilder returns null again.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        // The layout should not have changed.
+        expect(find.text('R0:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C5')),
+          const Rect.fromLTRB(200.0, 0.0, 400.0, 200.0),
+        );
+        expect(find.text('R4:C7'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C7')),
+          const Rect.fromLTRB(600.0, 800.0, 800.0, 1000.0),
+        );
+        // No columns laid out before column 3, but after column 7 we have added
+        // new columns.
+        expect(find.text('R0:C2'), findsNothing);
+        expect(find.text('R0:C8'), findsOneWidget);
+        // This exceeds the new bounds.
+        horizontalController.jumpTo(3200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected.
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 2400.0);
+        // Max scroll extent was updated to reflect reaching the end of the
+        // columns after returning null again at the new index.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, 2400.0);
+
+        // Decrease the number of columns
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            columnBuilder: (int index) {
+              // There will only be 5 columns.
+              if (index == 5) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+
+        // The position should have changed.
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 200.0);
+        // Max scroll extent was updated to the new end we have corrected to.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, 200.0);
+        // The layout updated.
+        expect(find.text('R0:C2'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C2')),
+          const Rect.fromLTRB(200.0, 0.0, 400.0, 200.0),
+        );
+        expect(find.text('R4:C4'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C4')),
+          const Rect.fromLTRB(600.0, 800.0, 800.0, 1000.0),
+        );
+        // No columns laid out after column 5.
+        expect(find.text('R0:C5'), findsNothing);
+      });
+
+      testWidgets('infinite columns can null terminate',
+          (WidgetTester tester) async {
+        // Nothing pinned ---
+        bool calledOutOfBounds = false;
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              if (index > 10) {
+                calledOutOfBounds = true;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.00);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated. This exceeds the bounds of the scroll view once the
+        // columns have been null terminated.
+        horizontalController.jumpTo(1400.0);
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        // Position was corrected.
+        expect(horizontalController.position.pixels, 1200.0);
+        // After null terminating, the builder was not called further.
+        expect(calledOutOfBounds, isFalse);
+        // Max scroll extent was updated to reflect reaching the end of the
+        // columns after returning null.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C9')),
+          const Rect.fromLTRB(600.0, 800.0, 800.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 9.
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C10'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            pinnedColumnCount: 1,
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.00);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated. This exceeds the bounds of the scroll view once the
+        // columns have been null terminated.
+        horizontalController.jumpTo(1400.0);
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        // Position was corrected.
+        expect(horizontalController.position.pixels, 1200.0);
+        // Max scroll extent was updated to reflect reaching the end of the
+        // columns after returning null.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C9')),
+          const Rect.fromLTRB(600.0, 800.0, 800.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 9, except for
+        // the pinned first column.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C10'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            pinnedRowCount: 1,
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.00);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated. This exceeds the bounds of the scroll view once the
+        // columns have been null terminated.
+        horizontalController.jumpTo(1400.0);
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        // Position was corrected.
+        expect(horizontalController.position.pixels, 1200.0);
+        // Max scroll extent was updated to reflect reaching the end of the
+        // columns after returning null.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C9')),
+          const Rect.fromLTRB(600.0, 800.0, 800.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 9.
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C10'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns and rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowCount: 10,
+            pinnedRowCount: 1,
+            pinnedColumnCount: 1,
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, 1400.00);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // Change the horizontal scroll offset, validate more columns were
+        // populated. This exceeds the bounds of the scroll view once the
+        // columns have been null terminated.
+        horizontalController.jumpTo(1400.0);
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        // Position was corrected.
+        expect(horizontalController.position.pixels, 1200.0);
+        // Max scroll extent was updated to reflect reaching the end of the
+        // columns after returning null.
+        expect(verticalController.position.maxScrollExtent, 1400.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C9')),
+          const Rect.fromLTRB(600.0, 800.0, 800.0, 1000.0),
+        );
+        // No columns laid out before column 5, or after column 9, except for
+        // the pinned first column.
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(find.text('R0:C4'), findsNothing);
+        expect(find.text('R0:C10'), findsNothing);
+      });
+      testWidgets('infinite rows & columns can null terminate',
+          (WidgetTester tester) async {
+        // Nothing pinned ---
+        bool calledRowOutOfBounds = false;
+        bool calledColumnOutOfBounds = false;
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              if (index > 8) {
+                calledRowOutOfBounds = true;
+              }
+              return largeSpan;
+            },
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              if (index > 10) {
+                calledColumnOutOfBounds = true;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the scroll offsets, validate more columns and rows were
+        // populated.
+        // These exceed the bounds of the scroll view once the column and row
+        // builders have null terminated.
+        horizontalController.jumpTo(1400.0);
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected for both.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        // After null terminating, the builders were not called further.
+        expect(calledRowOutOfBounds, isFalse);
+        expect(calledColumnOutOfBounds, isFalse);
+        // Max scroll extents were updated to reflect reaching the ends after
+        // returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C9')),
+          const Rect.fromLTRB(600.0, 400.0, 800.0, 600.0),
+        );
+        // No columns laid out before column 5, or after column 9.
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C10'), findsNothing);
+        // No rows laid out before row 4, or after row 7.
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R8:C6'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+            pinnedColumnCount: 1,
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the scroll offsets, validate more columns and rows were
+        // populated.
+        // These exceed the bounds of the scroll view once the column and row
+        // builders have null terminated.
+        horizontalController.jumpTo(1400.0);
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected for both.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        // Max scroll extents were updated to reflect reaching the ends after
+        // returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C9')),
+          const Rect.fromLTRB(600.0, 400.0, 800.0, 600.0),
+        );
+        // No columns laid out before column 5, or after column 9, except for
+        // the pinned first column.
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C10'), findsNothing);
+        // No rows laid out before row 4, or after row 7.
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R8:C6'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+            pinnedRowCount: 1,
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the scroll offsets, validate more columns and rows were
+        // populated.
+        // These exceed the bounds of the scroll view once the column and row
+        // builders have null terminated.
+        horizontalController.jumpTo(1400.0);
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected for both.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        // Max scroll extents were updated to reflect reaching the ends after
+        // returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C9')),
+          const Rect.fromLTRB(600.0, 400.0, 800.0, 600.0),
+        );
+        // No columns laid out before column 5, or after column 9.
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C10'), findsNothing);
+        // No rows laid out before row 4, or after row 7, except for first
+        // pinned row.
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R8:C6'), findsNothing);
+
+        await tester.pumpWidget(Container());
+
+        // Pinned columns and rows ---
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowBuilder: (int index) {
+              // There will only be 8 rows.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+            pinnedRowCount: 1,
+            pinnedColumnCount: 1,
+            columnBuilder: (int index) {
+              // There will only be 10 columns.
+              if (index == 10) {
+                return null;
+              }
+              return largeSpan;
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R4:C5'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R4:C5')),
+          const Rect.fromLTRB(1000.0, 800.0, 1200.0, 1000.0),
+        );
+        // No columns laid out beyond column 5.
+        expect(find.text('R0:C6'), findsNothing);
+        // No rows laid out beyond row 4.
+        expect(find.text('R5:C0'), findsNothing);
+        // Change the scroll offsets, validate more columns and rows were
+        // populated.
+        // These exceed the bounds of the scroll view once the column and row
+        // builders have null terminated.
+        horizontalController.jumpTo(1400.0);
+        verticalController.jumpTo(1200.0);
+        await tester.pumpAndSettle();
+        // Position was corrected for both.
+        expect(verticalController.position.pixels, 1000.0);
+        expect(horizontalController.position.pixels, 1200.0);
+        // Max scroll extents were updated to reflect reaching the ends after
+        // returning null.
+        expect(verticalController.position.maxScrollExtent, 1000.0);
+        expect(horizontalController.position.maxScrollExtent, 1200.0);
+        expect(find.text('R5:C6'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R5:C6')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 200.0),
+        );
+        expect(find.text('R7:C9'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R7:C9')),
+          const Rect.fromLTRB(600.0, 400.0, 800.0, 600.0),
+        );
+        // No columns laid out before column 5, or after column 9, except for
+        //the first pinned column.
+        expect(find.text('R5:C0'), findsOneWidget);
+        expect(find.text('R5:C4'), findsNothing);
+        expect(find.text('R5:C10'), findsNothing);
+        // No rows laid out before row 4, or after row 7, except for first
+        // pinned row.
+        expect(find.text('R0:C6'), findsOneWidget);
+        expect(find.text('R3:C6'), findsNothing);
+        expect(find.text('R8:C6'), findsNothing);
+      });
+      testWidgets('merged cells work with lazy layout computation',
+          (WidgetTester tester) async {
+        // When columns and rows are finite, the layout is eagerly computed and
+        // the children are lazily laid out. This makes computing merged cell
+        // layouts easy. In an infinite world, the layout is also lazily
+        // computed, so not all of the information may be available for a merged
+        // cell if it extends into an area we have not computed the layout for
+        // yet.
+        const ({int start, int span}) rowConfig = (start: 0, span: 10);
+        final List<int> mergedRows = List<int>.generate(
+          10,
+          (int index) => index,
+        );
+        const ({int start, int span}) columnConfig = (start: 1, span: 10);
+        final List<int> mergedColumns = List<int>.generate(
+          10,
+          (int index) => index + 1,
+        );
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            cellBuilder: (_, TableVicinity vicinity) {
+              // Merged row
+              if (mergedRows.contains(vicinity.row) && vicinity.column == 0) {
+                return TableViewCell(
+                  rowMergeStart: rowConfig.start,
+                  rowMergeSpan: rowConfig.span,
+                  child: const Text('R0:C0'),
+                );
+              }
+              // Merged column
+              if (mergedColumns.contains(vicinity.column) &&
+                  vicinity.row == 0) {
+                return TableViewCell(
+                  columnMergeStart: columnConfig.start,
+                  columnMergeSpan: columnConfig.span,
+                  child: const Text('R0:C1'),
+                );
+              }
+              return TableViewCell(
+                child: Text('R${vicinity.row}:C${vicinity.column}'),
+              );
+            },
+          ),
+        ));
+        await tester.pumpAndSettle();
+        expect(verticalController.position.pixels, 0.0);
+        expect(horizontalController.position.pixels, 0.0);
+        expect(verticalController.position.maxScrollExtent, double.infinity);
+        expect(horizontalController.position.maxScrollExtent, double.infinity);
+        expect(find.text('R0:C0'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C0')),
+          const Rect.fromLTRB(0.0, 0.0, 200.0, 2000.0),
+        );
+        expect(find.text('R0:C1'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R0:C1')),
+          const Rect.fromLTRB(200.0, 0.0, 2200.0, 200.0),
+        );
+        expect(find.text('R1:C1'), findsOneWidget);
+        expect(
+          tester.getRect(find.text('R1:C1')),
+          const Rect.fromLTRB(200.0, 200.0, 400.0, 400.0),
+        );
+      });
+
+      testWidgets('merged column that exceeds metrics will assert',
+          (WidgetTester tester) async {
+        final List<Object> exceptions = <Object>[];
+        final FlutterExceptionHandler? oldHandler = FlutterError.onError;
+        FlutterError.onError = (FlutterErrorDetails details) {
+          exceptions.add(details.exception);
+        };
+        const ({int start, int span}) columnConfig = (start: 1, span: 10);
+        final List<int> mergedColumns = List<int>.generate(
+          10,
+          (int index) => index + 1,
+        );
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            columnBuilder: (int index) {
+              // There will only be 8 columns, but the merge is set up for 10.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+            cellBuilder: (_, TableVicinity vicinity) {
+              // Merged column
+              if (mergedColumns.contains(vicinity.column) &&
+                  vicinity.row == 0) {
+                return TableViewCell(
+                  columnMergeStart: columnConfig.start,
+                  columnMergeSpan: columnConfig.span,
+                  child: const Text('R0:C1'),
+                );
+              }
+              return TableViewCell(
+                child: Text('R${vicinity.row}:C${vicinity.column}'),
+              );
+            },
+          ),
+        ));
+        await tester.pumpWidget(Container());
+        FlutterError.onError = oldHandler;
+        expect(exceptions.length, 3);
+        expect(
+          exceptions.first.toString(),
+          contains(
+            'The merged cell containing (row: 0, column: 1) is '
+            'missing TableSpan information necessary for layout. The '
+            'columnBuilder returned null, signifying the end, at column 8 but '
+            'the merged cell is configured to end with column 10.',
+          ),
+        );
+      });
+
+      testWidgets('merged row that exceeds metrics will assert',
+          (WidgetTester tester) async {
+        final List<Object> exceptions = <Object>[];
+        final FlutterExceptionHandler? oldHandler = FlutterError.onError;
+        FlutterError.onError = (FlutterErrorDetails details) {
+          exceptions.add(details.exception);
+        };
+        const ({int start, int span}) rowConfig = (start: 0, span: 10);
+        final List<int> mergedRows = List<int>.generate(
+          10,
+          (int index) => index,
+        );
+        await tester.pumpWidget(MaterialApp(
+          home: getTableView(
+            rowBuilder: (int index) {
+              // There will only be 8 rows, but the merge is set up for 9.
+              if (index == 8) {
+                return null;
+              }
+              return largeSpan;
+            },
+            cellBuilder: (_, TableVicinity vicinity) {
+              // Merged column
+              if (mergedRows.contains(vicinity.row) && vicinity.column == 0) {
+                return TableViewCell(
+                  rowMergeStart: rowConfig.start,
+                  rowMergeSpan: rowConfig.span,
+                  child: const Text('R0:C0'),
+                );
+              }
+              return TableViewCell(
+                child: Text('R${vicinity.row}:C${vicinity.column}'),
+              );
+            },
+          ),
+        ));
+        await tester.pumpWidget(Container());
+        FlutterError.onError = oldHandler;
+        expect(exceptions.length, 3);
+        expect(
+          exceptions.first.toString(),
+          contains(
+            'The merged cell containing (row: 0, column: 0) is '
+            'missing TableSpan information necessary for layout. The '
+            'rowBuilder returned null, signifying the end, at row 8 but '
+            'the merged cell is configured to end with row 9.',
+          ),
+        );
+      });
     });
   });
 
@@ -580,7 +2436,7 @@ void main() {
       await tester.pumpAndSettle();
 
       // Even columns and rows are set up for taps, mainAxis is vertical by
-      // default, mening row major order. Rows should take precedence where they
+      // default, meaning row major order. Rows should take precedence where they
       // intersect at even indices.
       expect(columnTapCounter, 0);
       expect(rowTapCounter, 0);
